% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.KLD_function.R
\name{barplot.KLD}
\alias{barplot.KLD}
\title{A barplot for the Kullback-Leibler divergence measure for all scenarios}
\usage{
\method{barplot}{KLD}(robust, compar, outcome, drug.names)
}
\description{
A barplot for the Kullback-Leibler divergence measure for all scenarios
}
