% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.model_function.R
\name{baseline_model}
\alias{baseline_model}
\title{The baseline model for binary outcome}
\usage{
baseline_model(base_risk, n_chains, n_iter, n_burnin, n_thin)
}
\arguments{
\item{base_risk}{Character string indicating the type of baseline model.
Set \code{base_type} equal to one of the following: \code{"fixed"},
\code{"random"}, or \code{"predicted"}.}

\item{n_chains}{Positive integer specifying the number of chains for the
MCMC sampling; an argument of the \code{\link[R2jags:jags]{jags}} function
of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 2.}

\item{n_iter}{Positive integer specifying the number of Markov chains for the
MCMC sampling; an argument of the \code{\link[R2jags:jags]{jags}} function
of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 10000.}

\item{n_burnin}{Positive integer specifying the number of iterations to
discard at the beginning of the MCMC sampling; an argument of the
\code{\link[R2jags:jags]{jags}} function of the R-package
\href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1000.}

\item{n_thin}{Positive integer specifying the thinning rate for the
MCMC sampling; an argument of the \code{\link[R2jags:jags]{jags}} function
of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1.}
}
\value{
An R character vector object to be passed to \code{\link{run_model}}
  and \code{\link{run_metareg}} through the
  \code{\link[base:textConnection]{textConnection}} function as the argument
  \code{object}.
}
\description{
XXXX
}
\details{
\code{prepare_model} creates the model in the JAGS dialect
  of the BUGS language. The output of this function constitutes the argument
  \code{model.file} of the \code{\link[R2jags:jags]{jags}} function (in the
  R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}) via the
  \code{\link[base:textConnection]{textConnection}} function.
}
\references{
Cooper NJ, Sutton AJ, Morris D, Ades AE, Welton NJ. Addressing between-study
heterogeneity and inconsistency in mixed treatment comparisons: Application
to stroke prevention treatments in individuals with non-rheumatic atrial
fibrillation. \emph{Stat Med} 2009;\bold{28}(14):1861--81.
doi: 10.1002/sim.3594

Dias S, Sutton AJ, Ades AE, Welton NJ. Evidence synthesis for decision
making 2: a generalized linear modeling framework for pairwise and network
meta-analysis of randomized controlled trials. \emph{Med Decis Making}
2013;\bold{33}(5):607--17. doi: 10.1177/0272989X12458724

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome
data in network meta-analysis: a one-stage pattern-mixture model approach.
\emph{Stat Methods Med Res} 2021;\bold{30}(4):958--75.
doi: 10.1177/0962280220983544

Spineli LM. An empirical comparison of Bayesian modelling strategies for
missing binary outcome data in network meta-analysis.
\emph{BMC Med Res Methodol} 2019;\bold{19}(1):86.
doi: 10.1186/s12874-019-0731-y

Turner NL, Dias S, Ades AE, Welton NJ. A Bayesian framework to account
for uncertainty due to missing binary outcome data in pairwise meta-analysis.
\emph{Stat Med} 2015;\bold{34}(12):2062--80. doi: 10.1002/sim.6475
}
\seealso{
\code{\link{run_metareg}}, \code{\link{run_model}},
   \code{\link[R2jags:jags]{jags}},
   \code{\link[base:textConnection]{textConnection}}
}
\author{
{Loukia M. Spineli}
}
