% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlandAltman.plot_function.R
\name{BlandAltman.plot}
\alias{BlandAltman.plot}
\title{Bland-Altman plot on the posterior mean of deviance between consistency and unrelated mean effects (UME) models}
\usage{
BlandAltman.plot(group1, group2, colour)
}
\description{
Bland-Altman plot on the posterior mean of deviance between consistency and unrelated mean effects (UME) models
}
