% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.clustering_function.R
\name{comp_clustering}
\alias{comp_clustering}
\title{End-user-ready results for informative and heuristic clustering of comparisons
(Comparisons' comparability for transitivity evaluation)}
\usage{
comp_clustering(
  input,
  drug_names,
  rename_char = NULL,
  height = FALSE,
  num_neighb,
  optimal_clusters,
  get_plots = "none",
  label_size = 4,
  title_size = 14,
  axis_title_size = 14,
  axis_text_size = 14,
  axis_x_text_angle = 0,
  legend_text_size = 13
)
}
\arguments{
\item{input}{A data-frame in the long arm-based format. Two-arm trials occupy
one row in the data-frame. Multi-arm trials occupy as many rows as the
number of possible comparisons among the interventions. The first three
columns refer to the trial name, first and second arm of the comparison,
respectively. The remaining columns refer to summary characteristics. See
'Details' for the specification of the columns.}

\item{drug_names}{A vector of labels with the name of the interventions
in the order they have been defined in the argument \code{input}.}

\item{rename_char}{A list of two elements: (i) a numeric vector with the
position of the characteristics in \code{input}, and (ii) a character
vector with the names of the characteristics, as they are wished to
appear in the title of the plots. This argument is optional, in case the
user wants to control the appearance of the titles.}

\item{height}{Logical with \code{TRUE} for performing heuristic clustering
and \code{FALSE} for performing heuristic clustering, thus, allowing the
user to define the number of clusters via the argument \code{optimal_clusters}.
The default argument is \code{FALSE}.}

\item{num_neighb}{A positive integer for the number of neighbouring
comparisons that is found in the \code{\link{connectivity_index}} and
\code{\link{internal_measures_plot}} functions. It takes values from two to
the number of comparisons minus one. The default argument equals half the
number of observed comparisons.}

\item{optimal_clusters}{A positive integer for the optimal number of clusters
based on three internal validation measures. The default argument is 2,
but the user \bold{must} adjust the value, if needed, after inspecting the
results of the internal validation measures. It takes values from two to
the number of comparisons minus one.}

\item{get_plots}{Logical with values \code{TRUE} for returning all plots and
\code{FALSE} for concealing the plots. The default argument is
\code{FALSE}.}

\item{label_size}{A positive integer for the font size of labels in the
bubble plot for the characteristics contribution, the violin plot for the
study dissimilarities per comparison, the bar plot for total dissimilarity
and the silhouette widths plot. \code{label_size} determines the size
argument found in the geom's aesthetic properties in the R-package
\href{https://CRAN.R-project.org/package=ggplot2}{ggplot2}).}

\item{title_size}{A positive integer for the font size of legend title in
the bubble plot for the characteristics contribution, the violin plot for
the study dissimilarities per comparison, the bar plot for total
dissimilarity and the silhouette widths plot. \code{title_size} determines
the title argument found in the theme's properties in the R-package
\href{https://CRAN.R-project.org/package=ggplot2}{ggplot2}).}

\item{axis_title_size}{A positive integer for the font size of axis title in
the bubble plot for the characteristics contribution, the violin plot for
the study dissimilarities per comparison, the bar plot for total
dissimilarity and the silhouette widths plot. \code{axis_title_size}
determines the axis.title argument found in the theme's properties in the
R-package \href{https://CRAN.R-project.org/package=ggplot2}{ggplot2}).}

\item{axis_text_size}{A positive integer for the font size of axis text in
the bubble plot for the characteristics contribution, the violin plot for
the study dissimilarities per comparison, the bar plot for total
dissimilarity and the silhouette widths plot. \code{axis_text_size}
determines the axis.text argument found in the theme's properties in the
R-package \href{https://CRAN.R-project.org/package=ggplot2}{ggplot2}).}

\item{axis_x_text_angle}{A positive integer for the angle of axis text in
the bubble plot for the characteristics contribution, the violin plot for
the study dissimilarities per comparison, and the bar plot for total
dissimilarity. \code{axis_x_text_angle} determines the axis.text.x argument
found in the theme's properties in the R-package
\href{https://CRAN.R-project.org/package=ggplot2}{ggplot2}).}

\item{legend_text_size}{A positive integer for the font size of legend text
in the bubble plot for the characteristics contribution, the violin plot
for the study dissimilarities per comparison, the bar plot for total
dissimilarity and the silhouette widths plot. \code{legend_text_size}
determines the legend.text argument found in the theme's properties in the
R-package \href{https://CRAN.R-project.org/package=ggplot2}{ggplot2}).}
}
\value{
Initially, \code{comp_clustering} prints on the console the following
  messages: the number of observed comparisons (and number of single-study
  comparisons, if any); the number of dropped characteristics due to many
  missing data; the maximum value of the cophenetic correlation coefficient;
  the optimal dissimilarity measure and optimal linkage method selected
  based on the cophenetic correlation coefficient. Then, the function
  returns the following list of elements:
  \item{Total_dissimilarity}{A data-frame on the total dissimilarity for each
  observed comparison. The data-frame has been sorted in decreasing order of
  the total dissimilarity.}
  \item{Types_used}{A data-frame with type mode (i.e., double or integer) of
  each characteristic.}
  \item{Total_missing}{The percentage of missing cases in the dataset,
  calculated as the ratio of total missing cases to the product of the number
  of studies with the number of characteristics.}
  \item{Cluster_color}{A data-frame on the number and color of the cluster
  assigned to each comparison. This can be used in the
  \code{\link{network_comparisons}} function to colour the edges according to
  the cluster assigned. The data-frame has been sorted in decreasing order of
  the total dissimilarity.}
  \item{Dissimilarity_table}{A lower off-diagonal matrix of 'dist' class
  with the dissimilarities of all pairs of comparisons. The row and column
  names has been sorted in decreasing order of the total dissimilarity.}
  \item{Table_internal_measures}{A data-frame with the connectivity index,
  silhouette width, and Dunn index for a range of 2 to P-1 clusters, with P
  being the number of comparisons.}
  \item{Table_cophenetic_coefficient}{A data-frame on the cophenetic
  correlation coefficient for all pairwise combinations of two dissimilarity
  measures (Euclidean, and Canberra) with eight linkage methods (Ward's two
  versions, single, complete, average, Mcquitty, median and centroid). The
  data-frame has been sorted in decreasing order of the cophenetic correlation
  coefficient.}
  \item{Optimal_dist}{The optimal dissimilarity measure (Euclidean or Canberra)
  based on the cophenetic correlation coefficient.}
  \item{Optimal_link}{The optimal linkage method (ward.D, ward.D2, single,
  complete, average, mcquitty, median, or centroid) based on the cophenetic
  correlation coefficient.}

  If \code{get_plots = FALSE} only the list of elements mentioned above is
  returned. If \code{get_plots = TRUE}, \code{comp_clustering} returns a
  series of plots in addition to the list of elements mentioned above:
  \item{Dissimilarity_comparison}{A violin plot with integrated box plots and
  dots on the study dissimilarities per comparison (x-axis). Violins are
  sorted in descending order of the total dissimilarity (red point).}
  \item{Characteristics_contribution}{A bubble plot on the percentage of
  average contribution of each characteristic (x-axis) to dissimilarities of
  pairs of studies in the corresponding comparison (y-axis).}
  \item{Total_dissimilarity_plot}{A barplot on the total dissimilarity of
  comparisons. Bars are sorted in descending order of the total dissimilarity
  and have been coloured based on the cluster they belong, with the clusters
  referring to the optimal partitioning determined by the argument
  \code{optimal_clusters}, when heuristic clustering is performed, or based
  on the informative clustering.}
  \item{Internal_measures_panel}{A panel of profile plots on the connectivity
  index, silhouette width, and Dunn index for a range of 2 to P-1 clusters,
  with P being the number of comparisons. The candidate optimal number of
  clusters is indicated with a red point directly on the line.}
  \item{Silhouette_comparisons}{A silhouette plot illustrating the silhouette
  width for each comparison. The comparisons are sorted in descending order
  of the silhouette width and have been coloured based on the cluster they
  belong, with the clusters referring to the optimal partitioning. For
  comparisons with zero silhouette width, 0.01 is added (an arbitrary very
  small number) to make visible the colour of the corresponding bars.}
}
\description{
\code{comp_clustering} hosts a toolkit of functions that facilitates
  conducting, visualising and evaluating informative and heuristic (hierarchical
  agglomerative) clustering of observed comparisons of interventions for a
  specific network and set of characteristics.
}
\details{
The correct type mode of columns in \code{input} must be ensured to use
  the function \code{comp_clustering}. The first three columns referring to
  the trial name, first and second arm of the comparison, respectively, must
  be \strong{character}. The remaining columns referring to the
  characteristics must be \strong{double} or \strong{integer} depending on
  whether the corresponding characteristic refers to a quantitative or
  qualitative variable. The type mode of each column is assessed by
  \code{comp_clustering} using the base function \code{typeof}. Note that
  \code{comp_clustering} invites unordered and ordered variables; for the
  latter, add the argument \code{ordered = TRUE} in the base function
  \bold{factor()}.

  The interventions should be sorted in an ascending order of their
  identifier number within the trials so that the first intervention column
  (second column in \code{input}) is the control arm for every pairwise
  comparison. This is important to ensure consistency in the intervention
  order within the comparisons obtained from the other related functions.

  \code{comp_clustering} excludes from the dataset the following type of
  characteristics: (i) completely missing characteristics and
  (ii) characteristics with missing values in all but one studies for at
  least one non-single-stufy comparison. Then it proceeds with the clustering
  process.

  The cophenetic correlation coefficient is calculated using the
  \code{\link[stats:cophenetic]{cophenetic}} function alongside the
  \code{\link[stats:dist]{dist}} function for selected dissimilarity measures
  found in the latter and the \code{\link[stats:hclust]{hclust}} function for
  selected linkage methods.

 \code{comp_clustering} can be used only for a network with at least three
  comparisons. Otherwise, the execution of the function will be stopped and
  an error message will be printed on the R console.
}
\references{
Gower J. General Coefficient of Similarity and Some of Its Properties.
\emph{Biometrics} 1971;\bold{27}(4):857--71.
doi: 10.2307/2528823

Sokal R, Rohlf F. The Comparison of Dendrograms by Objective Methods.
\emph{Int Assoc Plant Taxon} 1962;\bold{11}(2):33--40.
doi: 10.2307/1217208

Handl J, Knowles J, Kell DB. Computational cluster validation in post-genomic
data analysis. \emph{Biometrics} 2005;\bold{21}(15):3201--120.
doi: 10.1093/bioinformatics/bti517

Rousseeuw PJ. Silhouettes: A graphical aid to the interpretation and
validation of cluster analysis.
\emph{J Comput Appl Math} 1987;\bold{20}:53--65.

Dunn J. Well-separated clusters and optimal fuzzy partitions.
\emph{J Cybern} 1974;\bold{4}(1):95--104.
}
\seealso{
\code{\link{connectivity_index}},
 \code{\link[stats:cophenetic]{cophenetic}}, \code{\link[stats:dist]{dist}},
 \code{\link[stats:hclust]{hclust}}, \code{\link{internal_measures_plot}},
 \code{\link{network_comparisons}}
}
\author{
{Loukia M. Spineli}
}
