% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.index_function.R
\name{connectivity_index}
\alias{connectivity_index}
\title{Function for the Connectivity index
(Comparisons' comparability for transitivity evaluation)}
\usage{
connectivity_index(input, method, num_clusters, num_neighb)
}
\arguments{
\item{input}{An object of 'dist' class. It is a lower off-diagonal matrix
with the dissimilarities of all pairs of comparisons.}

\item{method}{A character string with values \code{"ward.D"},
\code{"ward.D2"}, \code{"single"}, \code{"complete"}, \code{"average"},
\code{"mcquitty"}, \code{"median"}, or \code{"centroid"} for the
linkage method.}

\item{num_clusters}{A positive integer for the number of clusters. It
takes values from two to the number of comparisons minus one.}

\item{num_neighb}{A positive integer for the number of neighbouring
comparisons. It takes values from two to the number of comparisons minus
one. The default argument equals half the number of comparisons.}
}
\value{
A scalar value for the connectivity index.
}
\description{
\code{connectivity_index} calculates the connectivity index for a specific
  linkage method, number of clusters and neighbouring comparisons.
}
\details{
\code{connectivity_index} is integrated in the function
  \code{\link{internal_measures_plot}}. The index ranges from zero to
  infinity with smaller values indicating higher connectivity for the
  selected number of clusters, which is desired.

  \code{connectivity_index} uses the \code{\link[stats:cutree]{cutree}}
  function (alongside the \code{\link[stats:hclust]{hclust}} function) to cut
  the dendrogram for a specified number of clusters via the argument
  \code{num_clusters} and linkage method via the argument \code{method}.
}
\references{
Handl J, Knowles J, Kell DB. Computational cluster validation in post-genomic
data analysis. \emph{Biometrics} 2005;\bold{21}(15):3201--120.
doi: 10.1093/bioinformatics/bti517
}
\seealso{
\code{\link[stats:cutree]{cutree}},
  \code{\link[stats:hclust]{hclust}}, \code{\link{internal_measures_plot}}
}
\author{
{Loukia M. Spineli}
}
