% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.preparation_function.R
\name{data.preparation}
\alias{data.preparation}
\title{Prepare the dataaet in the proper format for R2jags}
\usage{
data.preparation(data, measure)
}
\arguments{
\item{data}{The dataset}

\item{measure}{The effect measure}
}
\description{
Prepare the dataaet in the proper format for R2jags
}
