% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.heatmap_function.R
\name{dendro_heatmap}
\alias{dendro_heatmap}
\title{Dendrogram with amalgamated heatmap
(Comparisons' comparability for transitivity evaluation)}
\usage{
dendro_heatmap(input)
}
\arguments{
\item{input}{An object of S3 class \code{\link{comp_clustering}}. See 'Value'
in \code{\link{comp_clustering}}.}
}
\value{
\code{dendro_heatmap} uses the \code{\link[heatmaply:heatmaply]{heatmaply}}
  function of the R-package
  \href{https://CRAN.R-project.org/package=heatmaply}{heatmaply} to create a
  cluster heatmap for a selected linkage method and number of clusters. The
  function uses different colours to indicate the clusters directly on the
  dendrogram. The comparisons in the have been sorted in decreasing order of
  their total dissimilarity.

 @details
   The function inherits the dissimilarity measure (Euclidean or Canberra),
   linkage method and number of optimal clusters by the
   \code{\link{comp_clustering}} function.

   Remember: when using the \code{\link{comp_clustering}} function, inspect
   the results of  three internal measures (connectivity index, silhouette
   width, and Dunn index) for a wide range of clusters to decide on the optimal
   number of clusters.
}
\description{
\code{dendro_heatmap} creates a dendrogram alongside the heatmap of
  dissimilarities among the comparisons for a specific dissimilarity measure,
  linkage method and number of clusters.
}
\seealso{
\code{\link{comp_clustering}}, \code{\link[heatmaply:heatmaply]{heatmaply}}
}
\author{
{Loukia M. Spineli}
}
