% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug.names_function.R
\name{drug.names.alloc}
\alias{drug.names.alloc}
\title{Assign the original name of the interventions to their number id}
\usage{
drug.names.alloc(drug.names)
}
\arguments{
\item{drug.names}{An object of S3 class \code{run.sensitivity}.
Indicate all possible comparisons in the network.
The interventions in the drug.names should follow align with the id you considered for the function 'run.model'}
}
\description{
Assign the original name of the interventions to their number id
}
