% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn.index_function.R
\name{dunn_index}
\alias{dunn_index}
\title{Function for the Dunn index
(Comparisons' comparability for transitivity evaluation)}
\usage{
dunn_index(input, method, num_clusters)
}
\arguments{
\item{input}{An object of 'dist' class. It is a lower off-diagonal matrix
with the dissimilarities of all pairs of comparisons.}

\item{method}{A character string with values \code{"ward.D"},
\code{"ward.D2"}, \code{"single"}, \code{"complete"}, \code{"average"},
\code{"mcquitty"}, \code{"median"}, or \code{"centroid"} for the
linkage method.}

\item{num_clusters}{A positive integer for the number of clusters. It
takes values from two to the number of comparisons minus one.}
}
\value{
A scalar value for the Dunn index.
}
\description{
\code{dunn_index} calculates the Dunn index for a specific linkage method
  and number of clusters.
}
\details{
\code{dunn_index} is integrated in the function
  \code{\link{internal_measures_plot}}. The index ranges from zero to
  infinity, with values larger than one indicating good clustering.

  \code{dunn_index} uses the \code{\link[stats:cutree]{cutree}}
  function (alongside the \code{\link[stats:hclust]{hclust}} function) to cut
  the dendrogram for a specified number of clusters via the argument
  \code{num_clusters} and linkage method via the argument \code{method}.
}
\references{
Dunn J. Well-separated clusters and optimal fuzzy partitions.
\emph{J Cybern} 1974;\bold{4}(1):95--104.
}
\seealso{
\code{\link[stats:cutree]{cutree}},
  \code{\link[stats:hclust]{hclust}}, \code{\link{internal_measures_plot}}
}
\author{
{Loukia M. Spineli}
}
