% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.metareg_function.R
\name{forestplot.metareg}
\alias{forestplot.metareg}
\title{Comparator-specific forest-plot on estimated and predicted effect sizes}
\usage{
forestplot.metareg(full, reg, compar, cov.value, drug.names)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{reg}{An object of S3 class \code{\link{run.metareg}}. See 'Value' in \code{\link{run.metareg}}.}

\item{compar}{A character to indicate the comparator intervention. It must be any name found in \code{drug.names}.}

\item{cov.value}{A vector of two elements in the following order: a number that corresponds to a value of the covariate considered in \code{\link{run.metareg}},
and a character object to indicate the name of the covariate.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}
}
\value{
A panel of two forest plots: (1) a forest plot on the estimated effect size of comparisons with the selected intervention of the network, and
  (2) a forest plot predicted effect size of comparisons with the selected intervention of the network.
  Both panels illustrate the results from network meta-analysis and meta-regression using different colours.
}
\description{
This function illustrates a forest plot of the posterior mean and 95\% credible and predictive interval of comparisons with the selected intervention of the network.
}
\details{
The y-axis displays all interventions in the network; the selected intervention that comprises the \code{compar} is indicated in the plot with a homonymous label.
  The numerical results are displayed above each line.
  Odds ratio and ratio of means are reported in the original scale after exponentiation of the logarithmic scale.

  The interventions are sorted in the descending order of their SUCRA values obtain via network meta-analysis.

  \code{forestplot} can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed in the R console.
}
\references{
Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. \emph{J Clin Epidemiol} 2011;\bold{64}(2):163--71. [\doi{10.1016/j.jclinepi.2010.03.016}]
}
\seealso{
\code{\link{run.model}}, \code{\link{run.metareg}}
}
\author{
{Loukia M. Spineli}
}
