% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.panel.UME_function.R
\name{forestplot.panel.UME}
\alias{forestplot.panel.UME}
\title{A panel of forestplots on the posterior summaries of the observed comparisons in the network under the compared models}
\usage{
forestplot.panel.UME(full, ume, drug.names)
}
\arguments{
\item{ume}{A vector with the posterior standard deviation of the treatment effects of the observed comparisons obtained via the \code{UME.plot} function.}

\item{nma}{A vector with the posterior mean of the treatment effects of the observed comparisons obtained via the \code{UME.plot} function.}

\item{effect.size}{Text label on the effect measurd considered that appears on the x-axis.}

\item{expon}{Logical indicating whether the results should appear on the exponential scale. The default which is \code{expon = T}. It is relevant for
the odds ratio and ratio of means.}
}
\value{
A panel of forestplots on the posterior summaries of the observed comparisons in the network under the NMA, original and refined UME models.
The results refer to the posterior mean and 95% credible interval of the log OR. Grey panels refer to the omitted comparisons.
Red and green colours indicate weak and strong evidence, respectively. Namely, the corresponding 95% credible interval includes
and excludes the null value, respectively.
}
\description{
A panel of forestplots on the posterior summaries of the observed comparisons in the network under the compared models
}
