% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_function.R
\name{forestplot}
\alias{forestplot}
\title{Forest plot of NMA results with a selected intervention}
\usage{
forestplot(full, compar, drug.names)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{compar}{A character to indicate the comparator intervention. it must be any name found in \code{drug.names}.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the argument \code{drug.names} is not defined, the interventions are ordered as they appear in \code{data}.}
}
\value{
A panel of two forest plots: (1) a forest plot on the effect estimates and predictions of comparisons with the selected intervention in the network, and
(2) a forest plot on the posterior mean and 95\% credible interval of SUCRA values of the interventions (Salanti et al., 2011).
}
\description{
Provides a forest plot with the posterior mean and 95\% credible and prediction intervals for comparisons with the selected intervention in the network.
}
\details{
The y-axis of the plot displays the labels of the interventions in the network; the selected intervention that comprises the \code{compar} argument is annotated in the plot with the label 'Comparator intervention'.
  For each comparison with the selected intervention, the 95\% credible and prediction intervals are displayed as overlapping lines in different colours.
  The corresponding numerical results are displayed above each line: 95\% credible intervals are found in parentheses, and 95\% predictive intervals are found in brackets.
  Odds ratios and ratio of means are reported in the original scale after exponentiation of the logarithmic scale.

  The interventions are sorted in descending order of their SUCRA values.

  \code{forestplot} can be used only for a network of interventions.
}
\examples{
data("nma.liu2013")

# Show the first six trials of the dataset (one-trial-per-row format)
head(nma.liu2013)
#            study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
# Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA

\dontrun{
# Perform a random-effects network meta-analysis
res <- run.model(data = nma.liu2013,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter.prior = list("halfnormal", 0, 1),
                 mean.misspar = c(0, 0),
                 var.misspar = 1,
                 D = 1,
                 n.chains = 3,
                 n.iter = 10000,
                 n.burnin = 1000,
                 n.thin = 1)

# The names of the interventions in the order they appear in the dataset
interv.names <- c("placebo", "pramipexole", "serotonin–norepinephrine reuptake inhibitor",
                  "serotonin reuptake inhibitor", "tricyclic antidepressant", "pergolide")

# Create the league heatmap
forestplot(full = res, compar = "placebo", drug.names = interv.names)
}

}
\references{
Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. \emph{J Clin Epidemiol} 2011;\bold{64}(2):163--71. [\doi{10.1016/j.jclinepi.2010.03.016}]
}
\seealso{
\code{\link{run.model}}
}
\author{
{Loukia M. Spineli}
}
