% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.ref_function.R
\name{forestplot.ref}
\alias{forestplot.ref}
\title{Forest-plot of comparisons with the reference intervention}
\usage{
forestplot.ref(net, drug.names)
}
\description{
Forest-plot of comparisons with the reference intervention
}
