% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.observed.comparisons_function.R
\name{forestplot_comp}
\alias{forestplot_comp}
\title{Forest plot observed comparisons only}
\usage{
forestplot_comp(full, meta, node, drug_names, show)
}
\description{
Forest plot observed comparisons only
}
