% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.mod_function.R
\name{heatmap.mod}
\alias{heatmap.mod}
\title{Heatmap to illustrate the proportion of MOD in each intervention of every trial}
\usage{
heatmap.mod(data, trial.names, drug.names)
}
\description{
Heatmap to illustrate the proportion of MOD in each intervention of every trial
}
