% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.mod.trial_function.R
\name{heatmap.mod.trial}
\alias{heatmap.mod.trial}
\title{Heatmap of the risk of bias associated with missing participant outcome data in each trial-arm}
\usage{
heatmap.mod.trial(data, trial.names, drug.names)
}
\arguments{
\item{data}{A data-frame of a one-trial-per-row format containing arm-level data of each trial. This format is widely used for BUGS models.
See 'Format' in \code{\link[rnmamod]{run.model}} function for the specification of the columns.}

\item{trial.names}{A vector of labels with the name of the trials in the order they appear in the argument \code{data}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in the argument \code{data} is used, instead.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data}. If \code{drug.names} is not defined, the order of the interventions
as they appear in the argument \code{data} is used, instead.}
}
\value{
A heatmap presenting the percentage of MOD in each trial-arm of the dataset. The columns and the rows of the heatmap refer to the investigated interventions and trials, respectively.
  We used the 'five-and-twenty' rule of Sackett and colleagues to characterise the percentage of MOD as being associated with low (up to 5\%), moderate (more than 5\% and up to 20\%),
  and high risk of bias (more than 20\%). Low, moderate, high risk of bias due to MOD are indicated using green, orange, and red colour, respectively.
  The function is also applicable for a pairwise meta-analysis.
}
\description{
This function illustrates the risk of bias associated with missing participant outcome data (MOD) in each arm of every trial in the dataset.
}
\examples{
data("nma.liu2013")

# Return the first six trials of the dataset
head(nma.liu2013)
#>            study t1 t2 t3 r1 r2 r3 m1 m2 m3  n1  n2 n3
#>    Richard, 2012  1  3  4 15 16 23  6  8  4  39  42 34
#>     Barone, 2010  1  2 NA 27 38 NA 19 20 NA 152 144 NA
#> Weinbtraub, 2010  1  3 NA  2  5 NA  6  6 NA  27  28 NA
#>      Menza, 2009  1  4  5  4  2  9  6  7  5  17  18 17
#>      Devos, 2008  1  4  5  4  8 11  0  2  1  16  15 17
#>   Antonini, 2006  4  5 NA 10  8 NA  4  4 NA  16  15 NA

# The names of the interventions in the order they appear in the dataset
interv.names <- c("placebo", "pramipexole", "SNRI", "SSRI", "TCA", "pergolide")

# Create the heatmap
heatmap.mod.trial(data = nma.liu2013, trial.names = nma.liu2013$study, drug.names = interv.names)

}
\references{
Sackett DL, Richardson WS, Rosenberg WM, Haynes RB. Evidence-based medicine: how to practice and teach EBM.
New York: Churchill Livingstone 1997. ISBN: 0-443-05686-2.
}
\seealso{
\code{\link[rnmamod]{run.model}}
}
\author{
{Loukia M. Spineli}
}
