% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.outcome.dataset_function.R
\name{heatmap.outcome.dataset}
\alias{heatmap.outcome.dataset}
\title{Heatmap of the distribution of the outcome in the dataset}
\usage{
heatmap.outcome.dataset(data, trial.names, drug.names)
}
\arguments{
\item{data}{A data-frame of a one-trial-per-row format containing arm-level data of each trial. This format is widely used for BUGS models.
See 'Format' in \code{\link[rnmamod]{run.model}} function for the specification of the columns.}

\item{trial.names}{A vector of labels with the name of the trials in the order they appear in the argument \code{data}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data}. If \code{drug.names} is not defined, the order of the interventions
  as they appear in \code{data} is used, instead.

##CORRECT}
}
\value{
A heatmap presenting the outcome in each arm of every trial. IN case of a binary outcome, the heatmap illustrates the proportion of
  observed events in each arm of every trial. In the case of a continuous outcome, the heatmap illustrates the t-statistic in each arm of
  every trial. The trial-arm t-statistic is the ratio of the corresponding extracted mean outcome, \code{y}, to the standard error, \code{se} (see the \code{data.preparation} function).
The columns and the rows of the heatmap refer to the investigated interventions and trials, respectively.
}
\description{
This function illustrates the the distribution of the investigated outcome in each arm of every trial in the dataset.
}
\examples{
data("nma.schwingshackl2014")

# Return the first six trials of the dataset
head(nma.schwingshackl2014)
#          study t1 t2 t3    y1    y2 y3  sd1  sd2 sd3 m1 m2 m3 n1 n2 n3
#   Bacchi, 2012  1  2 NA -0.40 -0.35 NA 0.44 0.48  NA  1  1 NA 20 20 NA
#       Ku, 2010  1  2 NA -0.60 -0.30 NA 1.20 0.90  NA  0  0 NA 15 13 NA
#      Moe, 2011  1  2 NA -0.53 -0.35 NA 0.45 0.40  NA  1  2 NA 13 13 NA
#       Ng, 2010  1  2 NA -0.30 -0.40 NA 0.88 0.60  NA  0  0 NA 30 30 NA
#   Sukala, 2012  1  2 NA -0.10 -0.10 NA 0.51 0.93  NA  4  4 NA 13 13 NA
# Balducci, 2009  1  3 NA  6.34  6.65 NA 0.94 1.08  NA  0  0 NA 20 22 NA

# The names of the interventions in the order they appear in the dataset
interv.names <- c("aerobic", "resistance", "combined training")

# Create the heatmap
heatmap.outcome.dataset(data = nma.schwingshackl2014,
                        trial.names = nma.schwingshackl2014$study,
                        drug.names = interv.names)

}
\seealso{
\code{\link{run.model}}, \code{\link{data.preparation}}.
}
\author{
{Loukia M. Spineli}
}
