% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.robustness_function.R
\name{heatmap.robustness}
\alias{heatmap.robustness}
\title{Heatmap of robutness for NMA}
\usage{
heatmap.robustness(robust, drug.names, threshold)
}
\description{
Heatmap of robutness for NMA
}
