% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.similarity.UME_function.R
\name{heatmap.similarity.UME}
\alias{heatmap.similarity.UME}
\title{Heatmap of Kullback-Leibler divergence values: consistency model versus unrelated mean effects model}
\usage{
heatmap.similarity.UME(full, ume, drug.names, threshold)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{ume}{An object of S3 class \code{\link{run.UME}}. See 'Value' in \code{\link{run.UME}}.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}

\item{threshold}{A number indicating the threshold of similarity. See 'Details' below.}
}
\value{
A lower triangular heatmap matrix on the KLD measure in the summary effect size from the unrelated mean effects model
  to the consistency model (See, 'Description' in \code{similarity.index.UME}).
}
\description{
This functions facilitates the detection of observed pairwise comparisons that are sensible
  to applying the consistency model or the unrelated mean effects model. The heatmap is based on the Kullback-Leibler divergence (KLD) measure
  (Kullback and Leibler, 1951) in the summary effect size from the unrelated mean effects model to the consistency model (see \code{similarity.index.UME}).
}
\details{
\code{heatmap.similarity.UME} is integrated in the \code{UME.plot} function. The heatmap illustrates the KLD values for the observed pairwise comparisons in the network.
  The observed pairwise comparisons are read from left to right and are highlighted either with a green or red colour.
  Comparisons highlighted with green or red colour imply high or poor similarity of the posterior distribution of the corresponding summary effect size in the compared models.
  Cells with KLD values in white correspond to frail comparisons as detected by the \code{run.UME} function (see 'Details' in \code{run.UME}).

  The user may consider the values 0.28 and 0.17 as \code{threshold} for binary and continuous outcome data, respectively.
  These thresholds have been originally developed by Spineli et al. (2021) and considered also by Spineli (2021) in the proposed
  framework of global evaluation of the consistency assumption.
}
\references{
Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. \emph{submitted} 2021.

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. \emph{Res Synth Methods} 2021;\bold{12}(4):475--490. [\doi{10.1002/jrsm.1478}]

Kullback S, Leibler RA. On information and sufficiency. \emph{Ann Math Stat} 1951;\bold{22}(1):79--86. [\doi{10.1214/aoms/1177729694}]
}
\seealso{
\code{\link{run.model}}, \code{\link{run.UME}}, \code{\link{UME.plot}}, \code{\link{similarity.index.UME}}
}
\author{
{Loukia M. Spineli}
}
