% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.missing.network_function.R
\name{heatmap_missing_network}
\alias{heatmap_missing_network}
\title{Heatmap of the propoertion of missing participant outcome data in the network}
\usage{
heatmap_missing_network(data, drug_names)
}
\arguments{
\item{data}{A data-frame of a one-trial-per-row format containing arm-level
data of each trial. See 'Format' in \code{\link[rnmamod]{run_model}}
function for the specification of the columns.}

\item{drug_names}{A vector of labels with the name of the interventions in
the order they appear in the argument \code{data}. If the argument
\code{drug_names} is not defined, the interventions are ordered
as they appear in the argument \code{data}.}
}
\value{
A heatmap on the proportion of MOD in each intervention and observed
  comparison in the network. Each cell annotates the median, minimum and
  maximum (in parenthesis) proportion of MOD across the corresponding trials.
  The proportion of MOD in each intervention and observed comparison are
  depicted in white and black colour in the main diagonal and lower
  off-diagonal, of the heatmap, respectively, The pairwise comparisons are
  read from left to right.

  The 'five-and-twenty' rule of Sackett and colleagues is used to
  characterise the \bold{median} proportion of MOD as being associated with
  low (up to 5\%), moderate (more than 5\% and up to 20\%), and high risk of
  attrition bias (more than 20\%). Low, moderate, and high risk of bias
  associated with MOD is indicated using green, orange, and red colour,
  respectively.

  The summary statistics (median, minimum and maximum) for each intervention
  (main diagonal; white font) result from calculating the proportion of
  missing participants in each arm of every trial and then summarising across
  the corresponding trial-arms. Similarly, the summary statistics for each
  observed comparison (lower off-diagonal; black font) result from
  calculating the proportion of total missing participants in each trial and
  then summarising across the corresponding trials.

  The \code{heatmap_missing_network} function is redundant for a pairwise
  meta-analysis.
}
\description{
Illustrates the distribution of and the risk of bias associated
  with missing participant outcome data (MOD) for each intervention and
  observed comparison in the network.
}
\examples{
data("nma.stowe2011")

# Return the first six trials of the dataset
head(nma.stowe2011)
#              study t1 t2    y1    y2  sd1  sd2 m1 m2  n1  n2
#   DA (B): Interntl  1  2 -0.30 -1.20 4.36 4.32  7  3  83  84
#      DA (C): Spain  1  2 -2.47 -3.33 3.91 3.48  8  9  20  23
#         DA (C): UK  1  2 -0.70 -2.00 2.24 2.33  2  1  18  19
#      DA (C): USA 1  1  2 -0.77 -2.08 3.32 3.21 19 34  65 123
# DA (Pe): N America  1  2 -0.20 -1.80 4.79 4.81  0  0 187 189
# DA (Pr): CLEOPATRA  1  2 -0.90 -2.80 5.00 2.83  1  1 101 201

# The names of the interventions in the order they appear in the dataset
interv_names <- c("PBO+LD", "DA+LD", "COMTI+LD", "MAOBI+LD")

# Create the heatmap
heatmap_missing_network(data = nma.stowe2011,
                        drug_names = interv_names)

}
\references{
Sackett DL, Richardson WS, Rosenberg WM, Haynes RB. Evidence-based medicine:
how to practice and teach EBM. New York: Churchill Livingstone 1997.
ISBN: 0-443-05686-2.
}
\seealso{
\code{\link{run_model}}
}
\author{
{Loukia M. Spineli}
}
