% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.param.prior_function.R
\name{heterogeneity.param.prior}
\alias{heterogeneity.param.prior}
\title{Determine the prior distribution for the heterogeneity parameter}
\usage{
heterogeneity.param.prior(measure, model, heter.prior)
}
\arguments{
\item{measure}{The effect measure}

\item{model}{The analysis model}

\item{heter.prior}{The prior}
}
\description{
Determine the prior distribution for the heterogeneity parameter
}
