% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improved.UME_function.R
\name{improved.UME}
\alias{improved.UME}
\title{A function to detect the frail comparisons in multi-arm trials}
\usage{
improved.UME(t, N, ns, na)
}
\arguments{
\item{t}{A data-frame of the one-trial-per-row format containing the intervention identifier in each arm of every trial (see 'Details' below, and 'Arguments' in \code{run.model}).}

\item{N}{A data-frame of the one-trial-per-row format containing the number of participants randomised on the assigned intervention in each arm of every trial (see 'Details' below, and 'Arguments' in \code{run.model}).}

\item{ns}{A scale parameter on the number trials.}

\item{na}{A vector of length equal to \code{ns} with the number of arms in each trial.}
}
\value{
The output of \code{improved.UME} is a list of elements that are used by the \code{run.UME}:
\tabular{ll}{
 \code{nbase.multi} \tab A scalar parameter on the number of frail comparisons.\cr
 \tab \cr
 \code{t1.bn} \tab A vector with numeric values on the first arm of each frail comparison.\cr
 \tab \cr
 \code{t2.bn} \tab A vector with numeric values on the second arm of each frail comparison.\cr
 \tab \cr
 \code{base} \tab A vector with numeric values on the baseline intervention of the multi-arm trials that contain the frail comparisons.\cr
 \tab \cr
 \code{obs.comp} \tab A data-frame that indicates how many two-arms and multi-arm trials have included each pairwise comparison observed in the network.\cr
}
}
\description{
\code{improved.UME} detects the frail comparisons in multi-arm trials, that is, comparisons between non-baseline interventions not investigated in any two-arm trial of the network (Spineli, 2021).
  The 'original' model of Dias et al. (2013) omits the frail comparisons from the estimation process of the unrelated mean effects model.
  Consequently, their posterior distribution coincides with the prior distribution yielding implausible posterior standard deviations.
}
\details{
\code{improved.UME} is integrated in the \code{run.UME} function and calls the output of the \code{data.preparation} function after sorting the rows so that multi-arm trials appear at the bottom of the dataset.
  When there are no multi-arm trials or no frail comparisons in the network, \code{improved.UME} returns only the element \code{obs.comp} (see, 'Value').
}
\references{
Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. \emph{submitted} 2021.

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. \emph{Med Decis Making} 2013;\bold{33}(5):641--56. [\doi{10.1177/0272989X12455847}]
}
\seealso{
\code{\link{run.UME}}, \code{\link{data.preparation}}, \code{\link{run.model}}
}
\author{
{Loukia M. Spineli}
}
