% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalplot.panel.UME_function.R
\name{intervalplot.panel.UME}
\alias{intervalplot.panel.UME}
\title{A panel of interval plots: Consistency model versus unrelated mean effects model}
\usage{
intervalplot.panel.UME(full, ume, drug.names)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{ume}{An object of S3 class \code{\link{run.UME}}. See 'Value' in \code{\link{run.UME}}.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}.
If the argument \code{drug.names} is not defined, the order of the interventions as they appear in \code{data} is used, instead.}
}
\value{
A panel of interval plots on the posterior mean and 95\% credible interval of the summary effect size under the consistency model and the improved
  unrelated mean effects model (Spineli, 2021) of all pairwise comparisons observed in the network.
}
\description{
This function creates a panel of interval plots on the summary effect sizes under the consistency model and the unrelated mean effects model.
  The number of interval plots equals the number of pairwise comparisons observed in the network.
}
\details{
\code{intervalplot.panel.UME} is integrated in the \code{UME.plot} function. The consistency model and the unrelated mean effects model are
  abbreviated in the y-axis as 'NMA model' and 'UME model'. The intervals are highlighted with green, when the corresponding summary effect sizes do not cross the vertical line of no difference, and red otherwise.
  Grey panels refer to the frail comparisons as detected by the \code{run.UME} function (see 'Details' in \code{run.UME}).
}
\references{
Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. \emph{submitted} 2021.
}
\seealso{
\code{\link{run.model}}, \code{\link{run.UME}}, \code{\link{UME.plot}}
}
\author{
{Loukia M. Spineli}
}
