% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLD.barplot_function.R
\name{KLD.barplot}
\alias{KLD.barplot}
\title{A barplot for the Kullback-Leibler divergence measure for all scenarios}
\usage{
KLD.barplot(robust, compar, outcome, drug.names)
}
\description{
A barplot for the Kullback-Leibler divergence measure for all scenarios
}
