% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.heatmap_function.R
\name{league.heatmap}
\alias{league.heatmap}
\title{A heatmap league table with ordered interventions}
\usage{
league.heatmap(net, drug.names)
}
\arguments{
\item{net}{An object of S3 class \code{nma.continuous.full.model}.}

\item{drug.names}{A vector of characteristics with name of the interventions as appear in the function \code{nma.continuous.full.model}.}
}
\value{
A heatmap of the treatment effects of all possible comparisons in the off-diagonals, and the SUCRA values in the diagonals

\dontshow{load("netmodr/data/NMA_results.RData")}
}
\description{
A function to create a heatmap with the treatment effects of all possible comparisons of interventions.
The rows and columns of the heatmap refer to the interventions which are sorted by decreasing order from the best to the worst
by their SUCRA value. The off-diagonals contain the posterior mean and the 95\% credible interval
of the corresponding comparisons. The main diagonal contains the SUCRA values of the corresponding interventions. Results in the lower
triangle refer to comparisons in the opposite direction after converting negative values into positive values, and vice versa.
Darker shades of red correspond to larger treatment effects. Comparisons between interventions should be read from left to right and
the estimate in the cell refers to the row-defining intervention against the column-defining intervention. Results that indicate strong
evidence in favor of the row-defining intervention (i.e. the respective 95\% credible interval does not include the zero value of no difference)
are indicated with a double asterisk.
}
\examples{
drug.names <- sapply(1:14, function(x) letters[x])
league.heatmap(net = res, drug.names = drug.names)

}
