% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareg.plot_function.R
\name{metareg.plot}
\alias{metareg.plot}
\title{Plot the results from the meta-regression analysis}
\usage{
metareg.plot(full, reg, compar, cov.value, drug.names, save.xls)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{compar}{A character to indicate the comparator intervention. it must be any name found in \code{drug.names}.}

\item{cov.value}{A vector of two elements in the following order: a number that corresponds to a value of the covariate considered in \code{\link{run.metareg}},
and a character object to indicate the name of the covariate.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}

\item{save.xls}{Logical to indicate whether to export the tabulated results to an Excel 'xlsx' format (via the \code{\link[writexl]{write_xlsx}} function) to the working directory of the user.
The default is \code{FALSE} (do not export to an Excel format).}

\item{metareg}{An object of S3 class  \code{\link{run.metareg}}. See 'Value' in \code{\link{run.metareg}}.}
}
\description{
Plot the results from the meta-regression analysis
}
