% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareg.plot_function.R
\name{metareg_plot}
\alias{metareg_plot}
\title{End-user-ready results for network meta-regression analysis}
\usage{
metareg_plot(full, reg, compar, cov_value, drug_names, save_xls)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run_model}}. See 'Value' in
\code{\link{run_model}}.}

\item{reg}{An object of S3 class \code{\link{run_metareg}}. See 'Value' in
\code{\link{run_metareg}}.}

\item{compar}{A character to indicate the comparator intervention. It must be
any name found in \code{drug_names}.}

\item{cov_value}{A vector of two elements in the following order: a number
that corresponds to a value of the covariate considered in
\code{\link{run_metareg}}, and a character object to indicate the name of
the covariate. See also 'Details'.}

\item{drug_names}{A vector of labels with the name of the interventions in
the order they appear in the argument \code{data} of
\code{\link{run_model}}. If the argument \code{drug_names} is not defined,
the order of the interventions as they appear in \code{data} is used,
instead.}

\item{save_xls}{Logical to indicate whether to export the tabulated results
to an Excel 'xlsx' format (via the \code{\link[writexl]{write_xlsx}}
function) to the working directory of the user. The default is \code{FALSE}
(do not export to an Excel format).}
}
\value{
\code{metareg_plot} prints on the R console a message on the most
  parsimonious model (if any) based on the deviance information criterion
  (DIC; in red text). Furthermore, the function returns the following list of
   elements:
  \tabular{ll}{
   \code{table_estimates} \tab The posterior mean, and 95\% credible interval
   of the summary effect size for each comparison with the selected
   intervention under network meta-analysis and meta-regression.\cr
   \tab \cr
   \code{table_predictions} \tab The posterior mean, and 95\% predictive
   interval of the summary effect size for each comparison with the selected
   intervention under network meta-analysis and meta-regression.\cr
   \tab \cr
   \code{table_model_assessment} \tab The DIC, total residual deviance,
   number of effective parameters, and the posterior mean and 95% credible
   interval of between-trial standard deviation (\eqn{\tau}) under each model
   (Spiegelhalter et al. (2002)). When a fixed-effect model has been
   performed, \code{metareg_plot} does not return results on \eqn{\tau}.\cr
   \tab \cr
   \code{table_regression_coeffients} \tab The posterior mean and 95\%
   credible interval of the regression coefficient(s).\cr
   \tab \cr
   \code{Interval.plots} \tab The panel of forest-plots on estimated and
   predicted effect sizes of comparisons with the selected intervention under
   network meta-analysis and meta-regression. See 'Details' and 'Value' in
   \code{\link{forestplot_metareg}}.\cr
}
}
\description{
Illustrates the effect estimates, predictions and regression
  coefficients of comparisons with a specified comparator and also exports
  these results in an Excel format.
}
\details{
The DIC of the network meta-analysis model is compared with the DIC
  of the network meta-regression model. If the difference in DIC exceeds 5,
  the network meta-regression model is preferred; if the difference in DIC is
  less than -5, the network meta-analysis model is preferred; otherwise,
  there is little to choose between the compared models.

  When the covariate is binary, specify in the second element of
  \code{cov_value} the name of the level for which the scatterplot will be
  created.

  Furthermore, \code{metareg_plot} exports all tabulated results to separate
  Excel 'xlsx' formats (via the \code{\link[writexl]{write_xlsx}} function)
  to the working directory of the user.

  \code{metareg_plot} can be used only for a network of interventions. In the
  case of two interventions, the execution of the function will be stopped
  and an error message will be printed in the R console.
}
\examples{
data("nma.baker2009")

\dontrun{
# Perform a random-effects network meta-analysis
res <- run_model(data = nma.baker2009,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter_prior = list("halfnormal", 0, 1),
                 mean_misspar = c(0, 0),
                 var_misspar = 1,
                 D = 0,
                 n_chains = 3,
                 n_iter = 10000,
                 n_burnin = 1000,
                 n_thin = 1)

# Publicatiom year
pub_year <- c(1996, 1998, 1999, 2000, 2000, 2001, rep(2002, 5), 2003, 2003,
              rep(2005, 4), 2006, 2006, 2007, 2007)

# Perform a random-effects network meta-regression (exchangeable structure)
reg <- run_metareg(full = res,
                   covariate = pub_year,
                   covar_assumption = "exchangeable",
                   n_chains = 3,
                   n_iter = 10000,
                   n_burnin = 1000,
                   n_thin = 1)

# The names of the interventions in the order they appear in the dataset
interv_names <- c("placebo", "budesonide", "budesonide plus formoterol",
                  "fluticasone", "fluticasone plus salmeterol",
                  "formoterol", "salmeterol", "tiotropium")

# Plot the results from the network meta-analysis and meta-regression
# for comparisons with salmeterol and publication year 2000
metareg_plot(full = res,
             reg = reg,
             compar = "salmeterol",
             cov_value = c(2000, "publication year"),
             drug_names = interv_names)
}
}
\references{
Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries
for presenting results from multiple-treatment meta-analysis: an overview and
tutorial. \emph{J Clin Epidemiol} 2011;\bold{64}(2):163--71.
[\doi{10.1016/j.jclinepi.2010.03.016}]

Spiegelhalter DJ, Best NG, Carlin BP, van der Linde A. Bayesian measures of
model complexity and fit. \emph{J R Stat Soc B} 2002;\bold{64}:583--616.
[\doi{10.1111/1467-9868.00353}]
}
\seealso{
\code{\link{run_model}}, \code{\link{run_metareg}},
  \code{\link{forestplot_metareg}}
}
\author{
{Loukia M. Spineli}
}
