% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.param.prior_function.R
\name{missingness.param.prior}
\alias{missingness.param.prior}
\title{Indicate the mean value of the distribution of the missingness parameter}
\usage{
missingness.param.prior(assumption, mean.misspar)
}
\arguments{
\item{assumption}{The assumption about the IM}

\item{mean.misspar}{The mean value for the IM}
}
\description{
Indicate the mean value of the distribution of the missingness parameter
}
