% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.plot_function.R
\name{netplot}
\alias{netplot}
\title{The network plot}
\usage{
netplot(data, drug.names, ...)
}
\arguments{
\item{data}{A data-frame of a one-trial-per-row format containing arm-level data of each trial. This format is widely used for BUGS models.
See 'Format' in \code{run.model} for the specification of the columns.}

\item{drug.names}{A vector of characteristics with name of the interventions as appear in the function \code{run.model}.}
}
\value{
A network plot with coloured closed-loops informed by multi-arm trials.
}
\description{
A function to create the network plot. Each node refers to the intervention and each link refers to the observed pairwise comparison. 
Closed loops of interventions informed by multi-arm trials are indicated with different colours. The R-package \code{\link[pcnetmeta]}
is used to create the network plot.
}
