% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.comparisons_function.R
\name{network_comparisons}
\alias{network_comparisons}
\title{Network of pairwise comparisons of interventions
(Comparisons' comparability for transitivity evaluation)}
\usage{
network_comparisons(
  data,
  node_frame_color = "black",
  node_frame_width = 1,
  node_label_font = 1,
  node_label_cex = 1,
  node_label_dist = 0,
  node_label_color = "black",
  edge_level,
  edge_lty = 1,
  edge_curved = 0,
  ...
)
}
\arguments{
\item{data}{An object of S3 class \code{\link{comp_clustering}}. See 'Value'
in \code{\link{comp_clustering}}.}

\item{node_frame_color}{A character string for the colour of the node frames.
The default argument is "black". For more details refer to the R-package
\href{https://CRAN.R-project.org/package=igraph}{igraph}.}

\item{node_frame_width}{A positive integer for the size of the node frame.
The default argument is 1. For more details refer to the R-package
\href{https://CRAN.R-project.org/package=igraph}{igraph}.}

\item{node_label_font}{A positive integer for the font type of the node
labels. The default argument is 1. For more details refer to the R-package
\href{https://CRAN.R-project.org/package=igraph}{igraph}.}

\item{node_label_cex}{A positive integer for the size of the node labels.
The default argument is 1. For more details refer to the R-package
\href{https://CRAN.R-project.org/package=igraph}{igraph}.}

\item{node_label_dist}{A positive integer for the distance between label and
node. The default argument is 0. For more details refer to the R-package
\href{https://CRAN.R-project.org/package=igraph}{igraph}.}

\item{node_label_color}{A character string for the colour of the node labels.
The default argument is "black". For more details refer to the R-package
\href{https://CRAN.R-project.org/package=igraph}{igraph}.}

\item{edge_level}{A character string with values \code{"low"}, \code{"low"},
\code{"moderate"}, \code{"high"}, and \code{"very high"} to indicate the
level of dissimilarity between two comparisons to draw.
See 'Details' below.}

\item{edge_lty}{A positive integer for the line type of the edges. The
default argument is 1. For more details refer to the R-package
\href{https://CRAN.R-project.org/package=igraph}{igraph}.}

\item{edge_curved}{A value in the range 0 to 1 for the edge curvature. The
default argument is 0. For more details refer to the R-package
\href{https://CRAN.R-project.org/package=igraph}{igraph}.}

\item{...}{Further graphical arguments of the
\code{\link[igraph:plot.igraph]{plot.igraph}} function of the
R-package \href{https://CRAN.R-project.org/package=igraph}{igraph}}
}
\value{
The network of comparisons with different colours for the clustered nodes.
}
\description{
\code{network_comparisons} offers an alternative visualisation of the
  dendrogram with amalgamated heatmap. Presenting the comparisons in a
  circular layout, typically seen in networks, facilitates interpretation of
  the clustering results in the context of transitivity evaluation. It is
  relevant *only* for heuristic clustering.
}
\details{
The nodes refer to the observed comparisons and their size is proportional
   to the corresponding normalised total dissimilarity. The nodes are
   coloured with different colours to indicate the cluster they belong.
   The clusters and dissimilarity measure are inherited by
   \code{\link{comp_clustering}}. The edges refer to the dissimilarities of
   pairs of comparisons and their size is proportional to
   the corresponding normalised dissimilarity value. Four different colours
   are used to indicate the level of dissimilarity between two comparisons:
   green, yellow, orange, and red for low, moderate, high and very high
   dissimilarity (\eqn{d < 0.25}, \eqn{0.25 \qeq d < 0.5},
   \eqn{0.50 \qeq d < 0.75}, and \eqn{d \qeq 0.75}, respectively).

   Networks with many comparisons appear cluttered when showing all edges
   (i.e., \code{edge_level = "all"}). In this case, it is recommended to draw
   the network for the dissimilarity level of interest, for instance,
   \code{edge_level = "low"}). When a dissimilarity level does not exist, the
   network will be drawn without the corresponding edges.
}
\seealso{
\code{\link{comp_clustering}},
 \code{\link[igraph:plot.igraph]{plot.igraph}}
}
\author{
{Loukia M. Spineli}
}
