% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nma.baker2009}
\alias{nma.baker2009}
\title{Pharmacological interventions for chronic obstructive pulmonary disease}
\format{
A data-frame with 21 rows of arm-based data and 17 columns (maximum
   number of 4 arms).
}
\source{
Baker WL, Baker EL, Coleman CI. Pharmacologic treatments for chronic
obstructive pulmonary disease: a mixed-treatment
comparison meta-analysis. \emph{Pharmacotherapy} 2009;\bold{29}(8):891--905.
\doi{10.1592/phco.29.8.891}
}
\usage{
data(nma.baker2009)
}
\description{
A dataset of 21 trials comparing seven pharmacologic
  interventions with each other and placebo in chronic obstructive
  pulmonary disease (COPD) patients. The exacerbation of COPD is the analysed
  binary outcome.
}
\details{
The interventions have been coded as follows: 1, placebo; 2,
  budesonide; 3, budesonide plus formoterol; 4, fluticasone; 5, fluticasone
  plus salmeterol; 6, formoterol; 7, salmeterol; and 8, tiotropium.
}
\keyword{datasets}
