% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nma.bottomley2011}
\alias{nma.bottomley2011}
\title{Pharmacological interventions for moderately severe scalp psoriasis}
\format{
A data frame with 9 rows of arm-based data and 17 columns
  (maximum number of 4 arms).
}
\source{
Bottomley JM, Taylor RS, Ryttov J. The effectiveness of two-compound
formulation calcipotriol and betamethasone dipropionate gel in the
treatment of moderately severe scalp psoriasis: a systematic review of direct
and indirect evidence. \emph{Curr Med Res Opin} 2011;\bold{27}(1):251--268.
[\doi{10.1185/03007995.2010.541022}]
}
\usage{
data(nma.bottomley2011)
}
\description{
A dataset of 9 trials comparing six pharmacologic interventions
  with each other and placebo for moderately severe scalp psoriasis.
  The analysed binary outcome is the investigator's global assessment
  response at 4 weeks.
}
\details{
The interventions have been coded as follows: 1, betamethasone
  dipropionate; 2, betamethasone valerate; 3, calcipotriol; 4, calcipotriol
  plus polytar; 5, capasal; 6, two-compound formulation gel; and 7, placebo.
}
\keyword{datasets}
