% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nma.liu2013}
\alias{nma.liu2013}
\title{Antidepressants in Parkinson's disease}
\format{
A data frame with 11 rows of arm-based data and 13 columns (maximum
  number of 3 arms).
}
\source{
Liu J, Dong J, Wang L, Yan P, Sun S. Comparative efficacy and acceptability
of antidepressants in Parkinson's disease: a network meta-analysis.
\emph{PLoS One} 2013;\bold{8}(10):e76651.
doi: 10.1371/journal.pone.0076651
}
\usage{
data(nma.liu2013)
}
\description{
A dataset of 11 trials comparing the effectiveness of five
  antidepressants and placebo in Parkinson's disease. The analysed binary
  outcome is the number of patients with a reduction of at least 50/% from
  the baseline score.
}
\details{
The interventions have been coded as follows: 1, placebo;
  2, pramipexole; 3, serotonin-norepinephrine reuptake inhibitor;
  4, selective serotonin reuptake inhibitor; 5, tricyclic antidepressant; and
  6, pergolide.
}
\keyword{datasets}
