% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nma.malaria.donegan2018}
\alias{nma.malaria.donegan2018}
\title{Artemether, artesunate and quinine for severe malaria}
\format{
A data frame with 24 rows of arm-based data and 11 columns referring
  to the trial number, the treatment identifier in the compared arms,
  the odds ratio in the logarithmic scale and its standard error for each
  trial, the average age in years and its centered version, the number of
  events and number randomised for each arm and trial.
}
\source{
Donegan S, Dias S, Tudur-Smith C, Marinho V, Welton NJ. Graphs of study
contributions and covariate distributions for network meta-regression.
\emph{Res Synth Methods} 2018;\bold{9}(2):243--60. doi: 10.1002/jrsm.1292
}
\usage{
data(nma.malaria.donegan2018)
}
\description{
A dataset of 24 two-arm trials comparing artemether, artesunate
  and quinine with each other in patients with severe malaria. The analysed
  binary outcome is death.
}
\details{
The interventions have been coded as follows: 1, quinine;
  2, artemether; and 3, artesunate
}
\keyword{datasets}
