% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nma.schwingshackl2014}
\alias{nma.schwingshackl2014}
\title{Training modalities for patients with type 2 diabetes}
\format{
A data frame with 14 rows of arm-based data and 16 columns (maximum
  number of 3 arms).
}
\source{
Schwingshackl L, Missbach B, Dias S, et al. M Impact of different training
modalities on glycaemic control and blood lipids in patients with type 2
diabetes: a systematic review and network meta-analysis.
\emph{Diabetologia} 2014;\bold{57}(9):1789--97.
doi: 10.1007/s00125-014-3303-z
}
\usage{
data(nma.schwingshackl2014)
}
\description{
A dataset of 14 trials comparing three different training
  modalities (a triangle network) for patients with type 2 diabetes.
  The analysed continuous outcome is change from baseline in HbA1c levels.
}
\details{
The interventions have been coded as follows: 1, aerobic;
  2, resistance; and 3, combined training;
}
\keyword{datasets}
