% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodesplit.plot_function.R
\name{nodesplit.plot}
\alias{nodesplit.plot}
\title{Plot the results from the node-splitting approach}
\usage{
nodesplit.plot(node, full, drug.names)
}
\description{
Plot the results from the node-splitting approach
}
