% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pma.hetrick2012}
\alias{pma.hetrick2012}
\title{Paroxetine versus placeboe for depressive disorders}
\format{
A data frame with 4 rows of arm-based data and 9 columns.
}
\source{
Hetrick SE, McKenzie JE, Cox GR, Simmons MB, Merry SN. Newer generation
antidepressants for depressive disorders in children and adolescents.
\emph{Cochrane Database Syst Rev} 2012;\bold{11}:CD004851.
\doi{10.1002/14651858.CD004851.pub3}
}
\usage{
data(pma.hetrick2012)
}
\description{
A pairwise meta-analysis of 4 trials comparing paroxetine with
  placebo for depressive disorders in children and adolescents.
  The analysed binary outcome is remission or response as defined in trial.
}
\details{
The interventions have been coded as follows:
  1, placebo; 2, paroxetine.
}
\keyword{datasets}
