% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pma.taylor2004}
\alias{pma.taylor2004}
\title{Inositol versus glucose for depressive episode}
\format{
A data frame with 4 rows of arm-based data and 11 columns.
}
\source{
Taylor MJ, Wilder H, Bhagwagar Z, Geddes J. Inositol for depressive disorders. \emph{Cochrane Database Syst Rev} 2004;\bold{2}:CD004049. [\doi{10.1002/14651858.CD004049.pub2}]
}
\usage{
data(pma.taylor2004)
}
\description{
A pairwise meta-analysis of 4 trials comparing inositol with glucose (placebo) for depressive episode.
  The analysed continuous outcome is the resolution of a depressive episode using the Hamilton Depression Rating Scale.
}
\details{
The interventions have been coded as follows: 1, placebo; 2, inositol.
}
\keyword{datasets}
