% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.small.study.effect_function.R
\name{prepare.small.study.effect}
\alias{prepare.small.study.effect}
\title{Preparing the direction variable for the investigation of possible small-study effect}
\usage{
prepare.small.study.effect(data, measure, var.misspar, direction, D)
}
\arguments{
\item{data}{ddd}

\item{measure}{ddd}

\item{var.misspar}{ddd}

\item{direction}{A list referring to the direction of each intervention in the network}

\item{D}{A binary number for the direction of the outcome. Set \code{D = 1} for a positive outcome and \code{D = 0} for a negative outcome.}
}
\description{
Preparing the direction variable for the investigation of possible small-study effect
}
