% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow.similarities_function.R
\name{rainbow_similarities}
\alias{rainbow_similarities}
\title{Rainbow of Gower's similarity values for each study (Transitivity evaluation)}
\usage{
rainbow_similarities(
  results,
  axis_title_size = 12,
  axis_text_size = 12,
  label_size = 3.5
)
}
\arguments{
\item{results}{An object of S3 class \code{\link{comp_clustering}}.
See 'Value' in \code{\link{comp_clustering}}.}

\item{axis_title_size}{A positive integer for the font size of axis title
(both axes). \code{axis_title_size} determines the axis.title argument
found in the theme's properties in the R-package
\href{https://CRAN.R-project.org/package=ggplot2}{ggplot2}.}

\item{axis_text_size}{A positive integer for the font size of axis text (both
axes). \code{axis_text_size} determines the axis.text argument found in the
theme's properties in the R-package
\href{https://CRAN.R-project.org/package=ggplot2}{ggplot2}.}

\item{label_size}{A positive integer for the font size of labels appearing on
each study-specific segment. \code{label_size} determines the size argument
found in the geom's aesthetic properties in the R-package
\href{https://CRAN.R-project.org/package=ggplot2}{ggplot2}.}
}
\value{
A horizontal bar plot illustrating the range of Gower's similarity
values for each study using shades of red and green to indicate low and
substantial between-comparisons similarity, respectively: the darker the red,
the lower the between-comparisons similarity (corresponding to values close
to 0), whilst the darker the green, the higher the between-comparisons
similarity (corresponding to values close to 1). The study names appear on
the y-axis in the order they appear in \code{results} and the similarity
values appear on the x-axis. Red and blue points refer to the (average)
within-comparison and between-comparisons similarity, respectively, for each
study.
}
\description{
Illustrating the range of Gower's similarity values for each study in the
network.
}
\details{
The range of Gower's similarity values for each study result from calculating
the Gower's dissimilarity of a study versus the remaining studies in the
network for a set of clinical and methodological characteristics that may act
as effect modifiers. Then, the Gower's dissimilarities are transformed into
similarities by subtracting each value from 1: Gower's dissimilarities take
values from 0 to 1, with 0 and 1 implying perfect similarity and perfect
dissimilarity, respectively.

The exact similarity values appear as dotted, vertical, black lines on each
bar. The root mean square of the similarity values (fixed weights) also
appear on each bar as a black diamond.
}
\references{
Gower J. General Coefficient of Similarity and Some of Its Properties.
\emph{Biometrics} 1971;\bold{27}(4):857--71.
doi: 10.2307/2528823
}
\seealso{
\code{\link{comp_clustering}}
}
\author{
{Loukia M. Spineli}
}
