% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankosucra.plot_function.R
\name{rankosucra.plot}
\alias{rankosucra.plot}
\title{Integrated Rankograms and SUCRA Curves}
\usage{
rankosucra.plot(net, drug.names)
}
\arguments{
\item{net}{An object of S3 class \code{nma.continuous.full.model}.}

\item{drug.names}{A vector of characters with the names of the interventions in the order they appear in the function \code{nma.continuous.full.model}.}
}
\value{
Integrated rankograms and SUCRA curves for each intervention sorted in descending order by the SUCRA value.
}
\description{
Integrated Rankograms and SUCRA Curves
}
\examples{
drug.names <- sapply(1:14, function(x) letters[x])
rankogram.sucra.plots(net = res, drug.names = drug.names)

}
\references{
Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. J Clin Epidemiol. 2011;64(2):163-71.

\dontshow{load("netmodr/data/NMA_results.RData")}
}
