% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.nodesplit_function.R
\name{run.nodesplit}
\alias{run.nodesplit}
\title{A function to perform Bayesian node-splitting approach for aggregate binary or continuous outcomes}
\usage{
run.nodesplit(
  data,
  measure,
  assumption,
  heter.prior,
  net.ref,
  mean.misspar,
  var.misspar,
  n.chains,
  n.iter,
  n.burnin,
  n.thin
)
}
\description{
A function to perform Bayesian node-splitting approach for aggregate binary or continuous outcomes
}
