% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.separate.meta_function.R
\name{run.separate.meta}
\alias{run.separate.meta}
\title{A function to perform separate Bayesian random-effects pairwise meta-analyses for aggregate binary or continuous outcomes based on a connected network of interventions}
\usage{
run.separate.meta(data, net, n.chains, n.iter, n.burnin, n.thin)
}
\arguments{
\item{data}{A data-frame of a one-trial-per-row format containing arm-level data of each trial. This format is widely used for BUGS models. See 'Format' for the specification of the columns.}

\item{net}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{n.chains}{Integer specifying the number of chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 2.}

\item{n.iter}{Integer specifying the number of Markov chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 10000.}

\item{n.burnin}{Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1000.}

\item{n.thin}{Integer specifying the thinning rate for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1.}
}
\value{
An R2jags output on the summaries of the posterior distribution, and the Gelman–Rubin convergence diagnostic of the following parameters:
\describe{
 \item{\code{EM}}{The effect estimate of all observed comparisons of interventions in the network.}
 \item{\code{dev.o}}{The deviance contribution of each trial-arm based on the observed outcomes.}
 \item{\code{resdev.o}}{The total residual deviance contribution of every trial based on the observed outcomes.}
 \item{\code{totresdev.o}}{The total residual deviance based on the observed outcomes.}
 \item{\code{dev.m}}{The deviance contribution of each trial-arm based on the missing outcomes.}
 \item{\code{resdev.m}}{The total residual deviance contribution of every trial based on the missing outcomes.}
 \item{\code{resdev.m}}{The total residual deviance based on the missing outcomes.}
 \item{\code{tau}}{The between-trial standard deviation assumed to be common for all observed comparisons.}
}
}
\description{
A function to perform separate Bayesian random-effects pairwise meta-analyses for aggregate binary or continuous outcomes based on a connected network of interventions
}
\examples{
\dontshow{
load("./data/nma.baker2009.RData")
}

# Perform a random-effects NMA with consistency equations for the odds ratio (in the logarithmic scale) assuming missing at random for identical, intervention-specific informative missingness odds ratio.
res1 <- run.model(data = nma.baker2009, measure = "OR", model = "RE", assumption = "IDE-ARM", heter.prior = list("halfnormal", 0, 1), mean.misspar = 0, var.misspar = 1, D = 1, n.chains = 3, n.iter = 10000, n.burnin = 1000, n.thin = 1)

# Run separate random-effects pairwise meta-analyses using the same arguments with the network meta-analysis above.
run.separate.meta(data = data1, net = res1, n.chains = 3, n.iter = 10000, n.burnin = 1000, n.thin = 1)

}
\seealso{
\code{\link{R2jags}}
}
\author{
{Loukia M. Spineli}
}
