% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.series.meta_function.R
\name{run.series.meta}
\alias{run.series.meta}
\title{A series of Bayesian pairwise meta-analyses from a network of interventions}
\usage{
run.series.meta(full, n.chains, n.iter, n.burnin, n.thin)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{n.chains}{Integer specifying the number of chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 2.}

\item{n.iter}{Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 10000.}

\item{n.burnin}{Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1000.}

\item{n.thin}{Positive integer specifying the thinning rate for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1.}
}
\value{
An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:
\tabular{ll}{
 \code{EM} \tab The summary effect estimate of each pairwise comparison with at least two trials observed in the network.\cr
 \tab \cr
 \code{tau} \tab The between-trial standard deviation for pairwise comparisons with at least two trials, when the random-effects model has been specified.\cr
}
}
\description{
This function performs a Bayesian pairwise meta-analysis separately for pairwise comparisons with at least two trials observed in the investigated network of interventions.
}
\details{
\code{run.series.meta} does not contain the arguments \code{data}, \code{measure}, \code{model}, \code{assumption}, \code{heter.prior}, \code{mean.misspar}, and \code{var.misspar} that are found in \code{run.model}.
  This is to prevent misspecifying the Bayesian model as it would make the comparison of the consistency model (via \code{run.model}) with the separate pairwise meta-analyses for the observed comparisons meaningless.
  Instead, these arguments are contained in the argument \code{full} of the function. Therefore, the user needs first to apply \code{run.model}, and then use \code{run.series.meta} (see, 'Examples').

  \code{run.series.meta} runs Bayesian pairwise meta-analysis in \code{JAGS}. The progress of the simulation appears in the R console. The number of times \code{run.series.meta} is used appears in the R console as a text in red
  and it equals the number of pairwise comparisons observed in the network of interventions (see 'Examples').

  The output of \code{run.series.meta} is not end-user-ready. The \code{series.meta.plot} function uses the output of \code{run.series.meta} as an S3 object and processes it further to provide an end-user-ready output.

  \code{run.series.meta} can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed in the R console.
}
\examples{
data("nma.baker2009")

# Perform a random-effects network meta-analysis
res1 <- run.model(data = nma.baker2009,
                  measure = "OR",
                  model = "RE",
                  assumption = "IDE-ARM",
                  heter.prior = list("halfnormal", 0, 1),
                  mean.misspar = 0,
                  var.misspar = 1,
                  D = 1,
                  n.chains = 3,
                  n.iter = 10000,
                  n.burnin = 1000,
                  n.thin = 1)

# Run separate random-effects pairwise meta-analyses
run.series.meta(full = res1, n.chains = 3, n.iter = 10000, n.burnin = 1000, n.thin = 1)

\dontshow{
closeAllConnections()
}

}
\references{
Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. \emph{Stat Sci} 1992;\bold{7}:457--472. [\doi{10.1214/ss/1177011136}]
}
\seealso{
\code{\link{run.model}}, \code{\link{series.meta.plot}}, \code{\link[R2jags]{jags}}
}
\author{
{Loukia M. Spineli}
}
