% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.UME_function.R
\name{run.UME}
\alias{run.UME}
\title{A function to perform Bayesian unrelated mean effects model}
\usage{
run.UME(full, n.chains, n.iter, n.burnin, n.thin)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{n.chains}{Positive integer specifying the number of chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 2.}

\item{n.iter}{Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 10000.}

\item{n.burnin}{Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1000.}

\item{n.thin}{Positive integer specifying the thinning rate for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1.}
}
\value{
An R2jags output on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic of the following monitored parameters:
\tabular{ll}{
 \code{EM} \tab The summary effect estimate for each pairwise comparison observed in the network.\cr
 \tab \cr
 \code{EM.m} \tab The summary effect estimate for each pairwise comparison observed in multi-arm trials. Only multi-arm trials are considered for the estimation of the corresponding summary effects.\cr
 \tab \cr
 \code{dev.o} \tab The deviance contribution of each trial-arm based on the observed outcome.\cr
 \tab \cr
 \code{hat.par} \tab The fitted outcome at each trial-arm.\cr
 \tab \cr
 \code{tau} \tab The between-trial standard deviation (assumed common across the observed pairwise comparisons), when a random-effects model has been specified.\cr
}

The output also includes the following elements - the first three resulting from relevant monitored parameters:
\tabular{ll}{
 \code{leverage.o} \tab The leverage for the observed outcome at each trial-arm.\cr
 \tab \cr
 \code{sign.dev.o} \tab The sign of the difference between observed and fitted outcome at each trial-arm.\cr
 \tab \cr
 \code{model.assessment} \tab A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.\cr
 \tab \cr
 \code{jagsfit} \tab An object of S3 class \code{\link[R2jags]{jags}} with the posterior results on all monitored parameters to be used in the \code{mcmc.diagnostics} function.\cr
}
Furthermore, \code{run.UME} returns a character vector with the pairwise comparisons observed in the network, \code{obs.comp},
and a character vector with comparisons between non-baseline interventions observed in multi-arm trials only, \code{frail.comp}. Both vectors are used in \code{UME.plot} function.
}
\description{
This function performs a Bayesian network meta-analysis based on the unrelated mean effects model of Dias et al. (2013a) extended to address aggregate binary and continuous participant outcome data via the pattern-mixture model
  (Spineli, 2019; Spineli et al., 2021). This model offers a global evaluation of the plausibility of the consistency assumption in the network (Dias et al. (2013b)).
}
\details{
\code{run.UME} does not contain the arguments \code{data}, \code{measure}, \code{model}, \code{assumption}, \code{heter.prior}, \code{mean.misspar}, and \code{var.misspar} that are found in \code{run.model}.
  This is to prevent misspecifying the Bayesian model as it would make the comparison of the consistency model (via \code{run.model}) with the unrelated mean effects model meaningless.
  Instead, these arguments are contained in the argument \code{full} of the function. Therefore, the user needs first to apply \code{run.model}, and then use \code{run.UME} (see, 'Examples').

  Initially, \code{run.UME} calls the \code{improved.UME} function to identify the \emph{frail comparisons}, that is, comparisons between non-baseline interventions in multi-arm trials -- not investigated in any two-arm trial of the network (Spineli, 2021).
  The 'original' model of Dias et al. (2013a) omits the frail comparisons from the estimation process. Consequently, the number of estimated summary effect sizes is less than those obtained by performing separate pairwise meta-analyses (see \code{run.series.meta}).

  Then, to perform the Bayesian unrelated effect measures model, \code{run.UME} calls the \code{prepare.UME} function which contains the WinBUGS code as written by Dias et al. (2013a) for binomial and normal likelihood to analyse binary and continuous outcome data, respectively.
  \code{prepare.UME} has been extended to incorporate the pattern-mixture model with informative missingness parameters for binary and continuous outcome data (see, 'Details' in \code{run.model}).
  \code{prepare.UME} has been also 'upgraded' to account for the multi-arm trials by assigning conditional univariate normal distributions on the basic parameters of these trials, that is, effect parameters between non-baseline and baseline arms (Dias et al., 2013b; Spineli, 2021).

  \code{run.UME} runs Bayesian unrelated mean effects model in \code{JAGS}. The progress of the simulation appears in the R console.

  The output of \code{run.UME} is not end-user-ready. The \code{UME.plot} function uses the output of \code{run.UME} as an S3 object and processes it further to provide an end-user-ready output.

  \code{run.UME} can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed in the R console.
}
\examples{
data("nma.liu2013")

\dontrun{
# Perform a random-effects network meta-analysis (consistency model)
res <- run.model(data = nma.liu2013,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter.prior = list("halfnormal", 0, 1),
                 mean.misspar = 0,
                 var.misspar = 1,
                 D = 1,
                 n.chains = 3,
                 n.iter = 10000,
                 n.burnin = 1000,
                 n.thin = 1)

# Run random-effects unrelated mean effects model
run.UME(full = res, n.chains = 3, n.iter = 10000, n.burnin = 1000, n.thin = 1)
}

}
\references{
Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. \emph{submitted} 2021.

Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. \emph{Stat Methods Med Res} 2021. [\doi{10.1177/0962280220983544}]

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. \emph{BMC Med Res Methodol} 2019;\bold{19}(1):86. [\doi{10.1186/s12874-019-0731-y}]

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. \emph{Med Decis Making} 2013a;\bold{33}(5):641--56. [\doi{10.1177/0272989X12455847}]

Dias S, Sutton AJ, Ades AE, Welton NJ. Evidence synthesis for decision making 2: a generalized linear modeling framework for pairwise and network meta-analysis of randomized controlled trials. \emph{Med Decis Making} 2013b;\bold{33}(5):607--617. [\doi{10.1177/0272989X12458724}]

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. \emph{Stat Sci} 1992;\bold{7}:457--472. [\doi{10.1214/ss/1177011136}]
}
\seealso{
\code{\link{run.model}}, \code{\link{UME.plot}}, \code{\link{prepare.UME}}, \code{\link{run.series.meta}}, \code{\link[R2jags]{jags}}
}
\author{
{Loukia M. Spineli}
}
