% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.sucra_function.R
\name{scatteplot.sucra}
\alias{scatteplot.sucra}
\title{Scatterplot of surface under the cumulative ranking curve (SUCRA) values for two models}
\usage{
scatteplot.sucra(full, reg, cov.value, drug.names)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{reg}{An object of S3 class \code{\link{run.metareg}}. See 'Value' in \code{\link{run.metareg}}.}

\item{cov.value}{A vector of two elements in the following order: a number that corresponds to a value of the covariate considered in \code{\link{run.metareg}},
and a character object to indicate the name of the covariate.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}
}
\description{
XX
}
