% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.sucra_function.R
\name{scatterplot_sucra}
\alias{scatterplot_sucra}
\title{Scatterplot of SUCRA values}
\usage{
scatterplot_sucra(full, reg, cov_value, drug_names)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run_model}}.
See 'Value' in \code{\link{run_model}}.}

\item{reg}{An object of S3 class \code{\link{run_metareg}}. See 'Value' in
\code{\link{run_metareg}}.}

\item{cov_value}{A vector of two elements in the following order: a number
that corresponds to a value of the covariate considered in
\code{\link{run_metareg}}, and a character object to indicate the name of
the covariate. See also 'Details'.}

\item{drug_names}{A vector of labels with the name of the interventions in
the order they appear in the argument \code{data} of
\code{\link{run_model}}. If the argument \code{drug_names} is not defined,
the order of the interventions as they appear in \code{data} is used,
instead.}
}
\value{
A scatterplot of the SUCRA values under the network meta-analysis
  (y-axis) against the SUCRA values under the network meta-regression
  (x-axis) for a specific level or value of the investigated covariate.
  The names of the interventions appear at the corresponding points. Three
  coloured rectangulars appear in the scatterplot: a red rectangular refer to
  SUCRA values up to 50%, a yellow rectangular refer to SUCRA values between
  50% and 80%, and a green rectangular refers to SUCRA values over 80%.
  Interventions falling at the green area are considered as the highest
  ranked interventions, whilst interventions falling at the red area are
  considered as the lowest ranked interventions.
}
\description{
Creates a scatterplot of the SUCRA values under the network
  meta-analysis and the network meta-regression for a specific level or value
  of the investigated covariate.
}
\details{
The y-axis displays all interventions in the network; the selected
  intervention that comprises the \code{compar} is indicated in the plot with
  a homonymous label. The numerical results are displayed above each line.
  Odds ratio and ratio of means are reported in the original scale after
  exponentiation of the logarithmic scale.

  When the covariate is binary, specify in the second element of
  \code{cov_value} the name of the level for which the scatterplot will be
  created.

  \code{scatterplot_sucra} can be used only for a network of interventions. In
  the case of two interventions, the execution of the function will be
  stopped and an error message will be printed in the R console.
}
\references{
Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries
for presenting results from multiple-treatment meta-analysis: an overview and
tutorial. \emph{J Clin Epidemiol} 2011;\bold{64}(2):163--71.
[\doi{10.1016/j.jclinepi.2010.03.016}]
}
\seealso{
\code{\link{run_model}}, \code{\link{run_metareg}}
}
\author{
{Loukia M. Spineli}
}
