% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.deviance_function.R
\name{scatterplots_dev}
\alias{scatterplots_dev}
\title{Create the deviance scatterplot}
\usage{
scatterplots_dev(full, ume, colour)
}
\arguments{
\item{full}{A numeric vector with the posterior mean of deviance obtained
using the consistency model.}

\item{ume}{A numeric vector with the posterior mean of deviance obtained
using the unrelated mean effects model.}

\item{colour}{A string to define the colour of the points in the plot.}
}
\value{
A scatterplot on the posterior mean deviance contribution of the
  individual data points under the consistency model and the unrelated mean
  effects model. Each data point corresponds to a trial-arm indicated by a
  pair of numbers. The first number refers to the trial position in the
  dataset, and the second arm refers to the corresponding trial-arm
  (see 'Arguments' and 'Value' in \code{data_preparation}).
}
\description{
Illustrates the posterior mean of deviance of the individual
  data points under the unrelated mean effects model (via \code{run_ume})
  against the posterior mean of deviance under the consistency model
  (via \code{run_model}).
}
\details{
\code{scatterplots_dev} is integrated in the \code{ume_plot}
  function to create the scatterplot on the posterior mean of deviance under
  the compared models, also considered by Dias et al. (2013). When the
  majority of data points are scattered across the diagonal dotted line, we
  may conclude that the compared models have a good agreement. Data points
  scattered above and below the diagonal dotted line may contribute more to
  the poor fit of the unrelated mean effects model or the consistency model,
  respectively. However, using the \code{scatterplots_dev}, it is difficult
  to declare these data points as possible 'outliers'. Therefore,
  \code{scatterplots_dev} is accompanied by \code{bland_altman_plot} in the
  \code{ume_plot} function.
}
\references{
Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis
for decision making 4: inconsistency in networks of evidence based on
randomized controlled trials.
\emph{Med Decis Making} 2013a;\bold{33}(5):641--56.
[\doi{10.1177/0272989X12455847}]
}
\seealso{
\code{\link{ume_plot}}, \code{\link{run_model}},
  \code{\link{run_ume}}
}
\author{
{Loukia M. Spineli}
}
