% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.meta.plot_function.R
\name{separate.meta.plot}
\alias{separate.meta.plot}
\title{Plot the results from the unrelated mean effects model}
\usage{
\method{separate}{meta.plot}(full, meta, drug.names)
}
\description{
Plot the results from the unrelated mean effects model
}
