% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.meta.plot_function.R
\name{series_meta_plot}
\alias{series_meta_plot}
\title{End-user-ready results for series of pairwise meta-analyses}
\usage{
series_meta_plot(full, meta, drug_names, save_xls)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run_model}}. See 'Value' in
\code{\link{run_model}}.}

\item{meta}{An object of S3 class \code{\link{run_series_meta}}. See 'Value'
in \code{\link{run_series_meta}}.}

\item{drug_names}{A vector of labels with the name of the interventions in
the order they appear in the argument \code{data} of
\code{\link{run_model}}. If the argument \code{drug_names} is not defined,
the interventions are ordered as they appear in \code{data}.}

\item{save_xls}{Logical to indicate whether to export the tabulated results
in Excel 'xlsx' format at the working directory of the user.
  The default is \code{FALSE} (do not export in Excel format).}
}
\value{
\code{series_meta_plot} returns a panel of two forest plots: (1) a
  forest plot on the posterior mean and 95\% credible interval of the effect
  of the observed comparisons under the consistency model and the
  corresponding pairwise meta-analyses, and (2) a forest plot on the
  posterior median and 95\% credible interval of the between-trial standard
  deviation (\eqn{\tau}) for the observed comparisons. The estimated
  \eqn{\tau} from the consistency model appears as a rectangle in the forest
  plot. When a fixed-effect model has been fitted, only the forest plot on
  the estimated effects is shown.

  The R console prints the data-frame with the estimated effects of the
  observed comparisons under both models and the data-frame on the estimated
  (\eqn{\tau})s from the pairwise meta-analyses. Furthermore,
  \code{series_meta_plot} exports both data-frames in Excel 'xlsx' format at
  the working directory of the user.
}
\description{
Facilitates the comparison of the estimated effects and
  between-trial standard deviation (where applicable) from the network
  meta-analysis versus the results from the pairwise meta-analyses.
}
\details{
\code{series_meta_plot} can be used only for a network of
  interventions. The user can detect any inconsistencies in the estimated
  effects from the compared models and explore the gains in their precision
  by applying the consistency model. Furthermore, the user can investigate
  the plausibility of the common between-trial heterogeneity assumption which
  is typically considered in the consistency model.
}
\examples{
data("nma.dogliotti2014")

\dontrun{
# Perform a random-effects network meta-analysis
res <- run_model(data = nma.dogliotti2014,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter_prior = list("halfnormal", 0, 1),
                 mean_misspar = c(0, 0),
                 var_misspar = 1,
                 D = 1,
                 n_chains = 3,
                 n_iter = 10000,
                 n_burnin = 1000,
                 n_thin = 1)

# Run separate random-effects pairwise meta-analyses
meta <- run_series_meta(full = res,
                        n_chains = 3,
                        n_iter = 10000,
                        n_burnin = 1000,
                        n_thin = 1)

# The names of the interventions in the order they appear in the dataset
interv_names <- c("placebo", "aspirin", "aspirin plus clopidogrel",
                  "dabigatran 110 mg", "dabigatran 150 mg", "rivaroxaban",
                  "vitamin K antagonist", "apixaban")

# Plot the results from both models
series_meta_plot(full = res,
                 meta = meta,
                 drug_names = interv_names)
}
}
\seealso{
\code{\link{run_model}}, \code{\link{run_series_meta}}
}
\author{
{Loukia M. Spineli}
}
