% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.index_function.R
\name{silhouette_index}
\alias{silhouette_index}
\title{Function for the silhouette width and silhouette plot
(Comparisons' comparability for transitivity evaluation)}
\usage{
silhouette_index(input, method, num_clusters)
}
\arguments{
\item{input}{An object of 'dist' class. It is a lower off-diagonal matrix
with the dissimilarities of all pairs of comparisons.}

\item{method}{A character string with values \code{"ward.D"},
\code{"ward.D2"}, \code{"single"}, \code{"complete"}, \code{"average"},
\code{"mcquitty"}, \code{"median"}, or \code{"centroid"} for the
linkage method.}

\item{num_clusters}{A positive integer for the number of clusters. It
takes values from two to the number of comparisons minus one.}
}
\value{
\code{silhouette_index} returns the following list of elements:
  \item{silhoutte_comparisons}{A data-frame on the silhouette width of each
  comparison.}
  \item{silhoutte_width}{A scalar on the average silhouette width across the
  comparisons.}
}
\description{
\code{silhouette_index} calculates the average silhouette width and the
  comparison-specific silhouette widths for a specific linkage method and
  number of clusters.
}
\details{
\code{silhouette_index} is integrated in the function
  \code{\link{internal_measures_plot}}. Silhouette width ranges from minus to
  plus one, with values closer to 1 indicating higher compactness and
  separation for the selected clustering partition, whilst values closer to
  -1 indicate possible misclassification.

  \code{silhouette_index} uses the \code{\link[stats:cutree]{cutree}}
  function (alongside the \code{\link[stats:hclust]{hclust}} function) to cut
  the dendrogram for a specified number of clusters via the argument
  \code{num_clusters} and linkage method via the argument \code{method}.
}
\references{
Rousseeuw PJ. Silhouettes: A graphical aid to the interpretation and
validation of cluster analysis.
\emph{J Comput Appl Math} 1987;\bold{20}:53--65.
}
\seealso{
\code{\link[stats:cutree]{cutree}},
  \code{\link[stats:hclust]{hclust}}, \code{\link{internal_measures_plot}}
}
\author{
{Loukia M. Spineli}
}
