% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.similarity.UME_function.R
\name{similarity.index}
\alias{similarity.index}
\title{The heatmap of Kullback-Leibler divergence values of the observed comparisons under the network meta-analysis (NMA) and refined unrelated mean effects (UME) models}
\usage{
similarity.index(full, ume, threshold)
}
\arguments{
\item{threshold}{A number indicating the threshold of similarity.}

\item{robust}{The component \code{"robust"} of class \code{robustness.index}.}

\item{drug.names}{A vector of characteristics with name of the interventions as appear in the functions \code{run.model}, \code{run.UME}, or \code{run.UME.Dias}.}

\item{obs.comp}{A vector of the observed comparisons of interventions in the network obtained vias the \code{run.UME} function.}

\item{frail.comp}{A vector of the omitted comparisons of interventions among the observed comparisons in the network obtained vias the \code{run.UME} function.}
}
\value{
A heatmap on the Kullback-Leibler divergence (KLD) measure from the refined UME model to the NMA model for each observed comparison in the network.
The observed pairwise comparisons are read from left to right and are highlighted either with a green or red colour. Green and red colours imply high
and poor similarity of the posterior distribution of the treatment effects in the compared models. KLD values in white correspond to omitted comparisons.
}
\description{
The heatmap of Kullback-Leibler divergence values of the observed comparisons under the network meta-analysis (NMA) and refined unrelated mean effects (UME) models
}
