% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.index.UME_function.R
\name{similarity.index.UME}
\alias{similarity.index.UME}
\title{Similarity index: consistency model versus unrelated mean effects model}
\usage{
similarity.index.UME(full, ume, threshold)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{ume}{An object of S3 class \code{\link{run.UME}}. See 'Value' in \code{\link{run.UME}}.}

\item{threshold}{A number indicating the threshold of similarity. See 'Details' below.}
}
\value{
A list of the following two vectors:
\tabular{ll}{
 \code{KLD} \tab A numeric vector on the KLD values; one KLD value per observed comparison.\cr
 \tab \cr
 \code{robust} \tab A character vector on whether the divergence is acceptable or substantial.\cr
}

In \code{robust}, acceptable divergence is labeled as \emph{robust} (i.e., \code{KLD} \eqn{<} \code{threshold}), and substantial divergence is labeled as \emph{frail}. This is equivalent to concluding
that the summary effect size of a comparison is robust and sensible, respectively, to the applied model.
}
\description{
This function calculates the Kullback-Leibler divergence (KLD) measure in the summary effect size from the unrelated mean effects model
  to the consistency model. Then, for each pairwise comparison observed in the network, the KLD value is compared with a threshold that reflects
  the minimally allowed deviation between the compared models.
}
\details{
\code{similarity.index.UME} is integrated in the \code{UME.plot} function.
  The user may consider the values 0.28 and 0.17 as \code{threshold} for binary and continuous outcome data, respectively.
  These thresholds have been originally developed by Spineli et al. (2021) and considered also by Spineli (2021) in the proposed
  framework of global evaluation of the consistency assumption.
}
\references{
Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. \emph{submitted} 2021.

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. \emph{Res Synth Methods} 2021. [\doi{10.1002/jrsm.1478}]
}
\seealso{
\code{\link{run.model}}, \code{\link{run.UME}}, \code{\link{UME.plot}}
}
\author{
{Loukia M. Spineli}
}
