% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taylor.IMOR_function.R
\name{Taylor.IMOR}
\alias{Taylor.IMOR}
\title{Function to obtain logORs and SEs using the Taylor-series with IMOR}
\usage{
Taylor.IMOR(data, mean.value, var.value, rho)
}
\description{
Function to obtain logORs and SEs using the Taylor-series with IMOR
}
