% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taylor.IMDoM.IMRoM_function.R
\name{taylor_continuous}
\alias{taylor_continuous}
\title{Pattern-mixture model with Taylor series for continuous outcomes}
\format{
The columns of the data-frame in the argument \code{data} refer to
  the following ordered elements for a continuous outcome:
  \tabular{ll}{
   \strong{id} \tab A unique identifier for each trial.\cr
   \tab \cr
   \strong{y1} \tab The observed mean outcome in the first arm of the
   comparison.\cr
   \tab \cr
   \strong{y2} \tab The observed mean outcome in the second arm of the
   comparison.\cr
   \tab \cr
   \strong{sd1} \tab The observed standard deviation of the outcome in the
   first arm of the comparison.\cr
   \tab \cr
   \strong{sd2} \tab The observed standard deviation of the outcome in the
   second arm of the comparison.\cr
   \tab \cr
   \strong{m1} \tab The number of MOD in the first arm of the comparison.\cr
   \tab \cr
   \strong{m2} \tab The number of MOD in the second arm of the comparison.\cr
   \tab \cr
   \strong{c1} \tab The number of completers in the first arm of the
   comparison.\cr
   \tab \cr
   \strong{c2} \tab The number of completers in the second arm of the
   comparison.\cr
   \tab \cr
   \strong{t1} \tab An identified for the intervention in the first arm of
   the comparison.\cr
   \tab \cr
   \strong{t2} \tab An identified for the intervention in the second arm of
   the comparison.\cr
  }
}
\usage{
taylor_continuous(data, measure, mean_value, var_value, rho)
}
\arguments{
\item{data}{A data-frame in the long arm-based format. Two arm-trials occupy
one row in the data-frame. Multi-arm trials occupy as many rows as the
number of possible comparisons among the interventions. See 'Format' for
the specification of the columns.}

\item{measure}{Character string indicating the effect measure with values
\code{"MD"}, \code{"SMD"}, or \code{"ROM"} for the mean difference,
standardised mean difference and ratio of means, respectively.}

\item{mean_value}{A numeric value for the mean of the normal distribution of
the informative missingness parameter. The same value is considered for all
trial-arms of the dataset. The default argument is 0 and corresponds to the
missing-at-random assumption. For the informative missingness ratio of
means, the mean value is defined in the logarithmic scale.}

\item{var_value}{A positive non-zero number for the variance of the normal
distribution of the informative missingness parameter. When the
\code{measure} is \code{"MD"}, or \code{"SMD"} the default argument is 1;
When the \code{measure} is \code{"ROM"} the default argument is 0.04}

\item{rho}{A numeric value in the interval [-1, 1] that indicates the
correlation coefficient between two missingness parameters in a trial. The
same value is considered across all trials of the dataset. The default
argument is 0 and corresponds to uncorrelated missingness parameters.}
}
\value{
A data-frame that additionally includes the following elements:
  \tabular{ll}{
   \strong{EM} \tab The effect size adjusted for MOD and obtained using the
   Taylor series.\cr
   \tab \cr
   \strong{se.EM} \tab The standard error of the effect size for MOD and
   obtained using the Taylor series.\cr
  }
}
\description{
Applies pattern-mixture model under a specific assumption about
  the missingness parameter in trial-arms with \bold{continuous} missing
  participant outcome data (MOD) and uses the Taylor series to obtain the
  effect size and standard error for each trial (Mavridis et al., 2015).
}
\details{
The \code{taylor_continuous} function is found in the
  \code{\link[rnmamod]{unrelated_effects_plot}} function. The latter uses the
  the \code{\link[netmeta]{pairwise}} function from the package
  \href{https://CRAN.R-project.org/package=netmeta}{netmeta}
  to transform the dataset from the wide arm-based format
  (see, 'Arguments' for \code{data} in
  \code{\link[rnmamod]{unrelated_effects_plot}})
  into the long-arm based format.
}
\references{
Mavridis D, White IR, Higgins JP, Cipriani A, Salanti G. Allowing for
uncertainty due to missing continuous outcome data in pairwise and network
meta-analysis. \emph{Stat Med} 2015;\bold{34}(5):721--41.
[\doi{10.1002/sim.6365}]
}
\seealso{
\code{\link[rnmamod]{run_model}},
  \code{\link[rnmamod]{unrelated_effects_plot}},
  \code{\link[netmeta]{pairwise}}
}
\author{
{Loukia M. Spineli}
}
