% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UME.plot_function.R
\name{UME.plot}
\alias{UME.plot}
\title{Plot the results from the unrelated mean effects model}
\usage{
UME.plot(full, ume, drug.names, threshold)
}
\description{
Plot the results from the unrelated mean effects model
}
