% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloon.plot.mod_function.R
\name{balloon.plot.mod}
\alias{balloon.plot.mod}
\title{Enhanced balloon plot}
\usage{
balloon.plot.mod(sens, compar, D, drug.names)
}
\arguments{
\item{D}{A binary number for the direction of the outcome. Set \code{D = 1} for a positive outcome and \code{D = 0} for a negative outcome.}

\item{drug.names}{A vector of characters with the names of the interventions in the order they appear in the function \code{run.sensitivity}.}

\item{net}{An object of S3 class \code{run.sensitivity}.}
}
\description{
Enhanced balloon plot
}
