% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloon.plot.mod_function.R
\name{balloon.plot.mod}
\alias{balloon.plot.mod}
\title{Enhanced balloon plot: Investigate the impact of missing participant outcome data}
\usage{
balloon.plot.mod(sens, compar, drug.names)
}
\arguments{
\item{sens}{An object of S3 class \code{\link{run.sensitivity}}. See 'Value' in \code{\link{run.sensitivity}}.}

\item{compar}{A character vector with two characters that indicates the pairwise comparison of interest. The first character refers to the 'experimental' intervention
and the second character refers to the 'control' intervention of the comparison.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}
}
\value{
\code{balloon.plot.mod} A list of two enhanced balloon plots for one comparison (see 'Details' for the description of the enhanced balloon plot):
\tabular{ll}{
 \code{Plot.effect.size} \tab The enhanced balloon plot for the summary effect size of one pairwise comparison.\cr
 \tab \cr
 \code{Plot.tau} \tab The enhanced balloon plot for \eqn{\tau}. When the fixed-effect model has been performed in \code{run.sensitivity}, the function will not return the \code{Plot.tau}.\cr
}
}
\description{
This function creates the enhanced balloon plot for the summary effect size and between-trial standard deviation, \eqn{\tau}, under different scenarios about the missingness parameter
  for a pair of interventions. \code{balloon.plot.mod} uses the scenarios considered in \code{\link{run.sensitivity}}.
}
\details{
For the summary effect size of a pairwise comparison, the different colours and bubble sizes reflect the posterior standard deviation and the magnitude of the posterior mean, respectively.
  A colour key appears below the plot that matches the colours with the corresponding posterior standard deviation value.
  The size of the bubble is proportional to the magnitude of the corresponding posterior mean.
  Crossed bubbles indicate scenarios with conclusive evidence (i.e., the 95\% credible interval excludes the null value), and filled bubbles indicate scenarios with inconclusive evidence
  (i.e., the 95\% credible interval includes the null value).
  The missing-at-random assumption (the primary analysis) is labeled in a white frame.
  Both axes illustrate the scenarios as specified in the argument \code{mean.scenarios} of the \code{run.sensitivity} function:
  the x-axis refers to the 'experimental' intervention, and the y-axis refers to the 'control' intervention.

  The same enhanced balloon plot is created for \eqn{\tau}. However, filled bubbles indicate low extent of statistical heterogeneity,
  and crossed bubbles indicate considerable extent of statistical heterogeneity. Low extent implies that the posterior median of \eqn{\tau} is lower than
  the median of the prior distribution for the heterogeneity parameter (see, \code{\link{run.sensitivity}} and \code{\link{run.model}}); otherwise, there is considerable statistical heterogeneity.

  \code{heatmap.robustness} can be used only for a network of interventions and when missing participant outcome data have been extracted for at least one trial.
  Otherwise, the execution of the function will be stopped and an error message will be printed in the R console.
}
\examples{
data("nma.baker2009")

# Perform a random-effects network meta-analysis
res1 <- run.model(data = nma.baker2009,
                  measure = "OR",
                  model = "RE",
                  assumption = "IDE-ARM",
                  heter.prior = list("halfnormal", 0, 1),
                  mean.misspar = 0,
                  var.misspar = 1,
                  D = 1,
                  n.chains = 3,
                  n.iter = 10000,
                  n.burnin = 1000,
                  n.thin = 1)

# Perform the sensitivity analysis (using the 'default' of the argument 'mean.scenarios')
res.sens <- run.sensitivity(full = res1,
                            assumption = "IDE-ARM",
                            var.misspar = 1,
                            n.chains = 3,
                            n.iter = 10000,
                            n.burnin = 1000,
                            n.thin = 1)

# The names of the interventions in the order they appear in the dataset
interv.names <- c("budesodine", "budesodine plus formoterol", "fluticasone", "fluticasone plus
                  salmeterol", "formoterol", "salmeterol", "tiotropium", "placebo")

# Create the enhanced balloon plot for the comparison 'tiotropium versus salmeterol'
balloon.plot.mod(sens = res.sens, compar = c("tiotropium", "salmeterol"), drug.names = interv.names)

}
\references{
Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. \emph{Res Synth Methods} 2021;\bold{12}(4):475--490. [\doi{10.1002/jrsm.1478}]
}
\seealso{
\code{\link{run.sensitivity}}, \code{\link{run.model}}
}
\author{
{Loukia M. Spineli}
}
