% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloon.plot.mod_function.R
\name{balloon_plot}
\alias{balloon_plot}
\title{Enhanced balloon plot}
\usage{
balloon_plot(sens, compar, drug_names)
}
\arguments{
\item{sens}{An object of S3 class \code{\link{run_sensitivity}}.
See 'Value' in \code{\link{run_sensitivity}}.}

\item{compar}{A character vector with two elements indicating the pairwise
comparison of interest. The first element refers to the 'experimental'
and the second element to the 'control' intervention of the comparison.}

\item{drug_names}{A vector of labels with the name of the interventions
in the order they appear in the argument \code{data} of
\code{\link{run_model}}. If the argument \code{drug_names} is not defined,
the order of the interventions as they appear in
\code{data} is used, instead.}
}
\value{
\code{balloon_plot} returns two enhanced balloon plots for one
  comparison (see 'Details'):
  \tabular{ll}{
   \code{Plot_effect_size} \tab The enhanced balloon plot for the
   summary effect size (according to the argument \code{measure} inherited
   by \code{\link{run_sensitivity}}) of one pairwise comparison.\cr
   \tab \cr
   \code{Plot_tau} \tab The enhanced balloon plot for tau. When the
   fixed-effect model has been performed in \code{\link{run_sensitivity}},
   the function will not return the \code{Plot_tau}.\cr
}
}
\description{
Creates the enhanced balloon plot for the summary effect size and
  between-trial standard deviation, \emph{tau}, under different scenarios
  about the missingness parameter for a pair of interventions.
  \code{balloon_plot} uses the scenarios considered in
  \code{\link{run_sensitivity}}.
}
\details{
For the \code{Plot_effect_size} of the selected pairwise
  comparison, the different colours and sizes of the bubbles reflect the
  posterior standard deviation and the posterior mean, respectively.
  A colour key appears below the plot.
  The size of the bubble is proportional to the corresponding posterior mean.
  Crossed bubbles indicate scenarios with conclusive evidence (the
  95\% credible interval excludes the null value), and filled bubbles
  indicate scenarios with inconclusive evidence (the 95\% credible interval
  includes the null value). The missing-at-random assumption (primary
  analysis) is labeled in a white frame.
  Both axes illustrate the scenarios as specified in the argument
  \code{mean_scenarios} of the \code{\link{run_sensitivity}}:
  the x-axis refers to the 'experimental' intervention, and the y-axis refers
  to the 'control' intervention.

  The same enhanced balloon plot is created for tau (\code{Plot_tau}).
  However, filled bubbles indicate low statistical heterogeneity
  (the posterior median of tau is lower than the median of the
  prior distribution for the heterogeneity parameter),
  and crossed bubbles indicate considerable statistical heterogeneity
  (the posterior median of tau exceeds the median of the prior
  distribution for the heterogeneity parameter).

  \code{balloon_plot_mod} can be used only when missing participant
  outcome data have been extracted for at least one trial.
  Otherwise, the execution of the function will be stopped and an error
  message will be printed on the R console.
}
\examples{
data("nma.baker2009")

\dontrun{
# Perform a random-effects network meta-analysis
res <- run_model(data = nma.baker2009,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter_prior = list("halfnormal", 0, 1),
                 mean_misspar = c(0, 0),
                 var_misspar = 1,
                 D = 1,
                 n_chains = 3,
                 n_iter = 10000,
                 n_burnin = 1000,
                 n_thin = 1)

# Perform the sensitivity analysis (default values for 'mean_misspar')
res_sens <- run_sensitivity(full = res,
                            assumption = "IDE-ARM",
                            var_misspar = 1,
                            n_chains = 3,
                            n_iter = 10000,
                            n_burnin = 1000,
                            n_thin = 1)

# The names of the interventions in the order they appear in the dataset
interv_names <- c("placebo", "budesonide", "budesonide plus formoterol",
                  "fluticasone", "fluticasone plus salmeterol",
                  "formoterol", "salmeterol", "tiotropium")

# Create the enhanced balloon plot for 'tiotropium versus salmeterol'
balloon_plot(sens = res_sens,
             compar = c("tiotropium", "salmeterol"),
             drug_names = interv_names)
}
}
\references{
Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of
primary analysis results: A case study on missing outcome data in pairwise
and network meta-analysis. \emph{Res Synth Methods}
2021;\bold{12}(4):475--490. \doi{10.1002/jrsm.1478}
}
\seealso{
\code{\link{run_model}}, \code{\link{run_sensitivity}}
}
\author{
{Loukia M. Spineli}
}
