% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.preparation_function.R
\name{data.preparation}
\alias{data.preparation}
\title{Prepare the dataset in the proper format for R2jags}
\usage{
data.preparation(data, measure)
}
\arguments{
\item{data}{A data-frame of the one-trial-per-row format containing arm-level data of each trial. This format is widely used for BUGS models. See 'Format' in the \code{run.model} function.}

\item{measure}{Character string indicating the effect measure with values \code{"OR"}, \code{"MD"}, \code{"SMD"}, or \code{"ROM"} for the odds ratio, mean difference,
standardised mean difference and ratio of means, respectively.}
}
\value{
A list of data-frames on the following elements to be passed to \code{\link{run.model}}:
\tabular{ll}{
 \code{m} \tab The number of missing participant outcome data in each arm of every trial (see 'Details' in the \code{run.model} function).\cr
 \tab \cr
 \code{N} \tab The number of participants randomised on the assigned intervention in each arm.\cr
 \tab \cr
 \code{t} \tab The intervention identifier in each arm of every trial.\cr
 \tab \cr
 \code{I} \tab The pseudo-data-frame \code{I} (see 'Details' in the \code{run.model} function).\cr
 \tab \cr
 \code{r} \tab The observed number of events of the outcome in each arm of every trial, when the outcome is binary.\cr
 \tab \cr
 \code{y0} \tab The observed mean value of the outcome in each arm, when the outcome is continuous\cr
 \tab \cr
 \code{se0} \tab The observed standard deviation of the outcome in each arm, when the outcome is continuous\cr
}
}
\description{
This function prepares the dataset in the proper format for R2jags. \code{data.preparation} is found in the \code{run.model} function; thus, the arguments are as specified in the latter.
}
\details{
\code{data.preparation} prepares the data for the Bayesian analysis. The data preparation includes the following actions. First, \code{data.preparation} checks
  whether the element \strong{m} exists in \code{data}. If this element is missing, \code{data.preparation} creates a pseudo-data-frame for \code{m} that has the zero value for the observed trial-arms, and \code{NA} for the unobserved trial-arms,
  and the pseudo-data-frame \code{I} that has the same values with the pseudo-data-frame for \code{m}. If the element \strong{m} exists in \code{data} and has values only for some trials, the pseudo-data-frame for \code{m} has the same values with
  \code{m} for the corresponding trial-arms, and the pseudo-data-frame \code{I} has the value one for these trial-arms. Both pseudo-data-frames aim to retain the trials without information on MOD in \code{data} and allow running the Bayesian analysis
  when MOD have not been extracted for any trial of the dataset. Second, \code{data.preparation} sorts the interventions across the arms of each trial in an ascending order and correspondingly sorts within each trial the remaining elements in \code{data}
  (see also 'Format' in \code{run.model}). Since all Bayesian models in the package consider the first column in \strong{t} to include the control arm for every trial, this sorting ensures that interventions with a lower identifier are consistently treated as the control arm in each trial. This case is relevant in
  non-star-shaped networks.
}
\seealso{
\href{https://CRAN.R-project.org/package=R2jags}{R2jags}, \code{\link{run.model}}
}
\author{
{Loukia M. Spineli}
}
