% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.network_function.R
\name{describe_network}
\alias{describe_network}
\title{A function to describe a network of interventions}
\usage{
describe_network(data, drug_names, measure)
}
\arguments{
\item{data}{A data-frame of a one-trial-per-row format containing arm-level
data of each trial. This format is widely used for BUGS models.
See 'Format' in \code{\link[rnmamod]{run_model}} function for the
specification of the columns.}

\item{drug_names}{A vector of labels with the name of the interventions in
the order they appear in the argument \code{data}. If the argument
\code{drug_names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}

\item{measure}{Character string indicating the effect measure with values
\code{"OR"}, \code{"MD"}, \code{"SMD"}, or \code{"ROM"} for the odds ratio,
mean difference, standardised mean difference and ratio of means,
respectively.}
}
\value{
A list of scalar results and four data-frames to be passed to
  \code{\link[rnmamod]{netplot}}. The scalar results include:
  \tabular{ll}{
   \code{direct_comp} \tab The number of observed comparisons in the
   network. \cr
   \tab \cr
   \code{two_arm_ns} \tab The number of two-arm trials in the network. \cr
   \tab \cr
   \code{multi_arm_ns} \tab The number of multi-arm trials in the
   network. \cr
   \tab \cr
   \code{total_rand_network} \tab The total number of randomised participants
    in the network. \cr
   \tab \cr
   \code{prop_obs_network} \tab The proportion of participants who completed
    the trial. \cr
   \tab \cr
   \code{prop_event_network} \tab The proportion of observed events in the
   network. When the outcome is continuous, this element is omitted. \cr
   \tab \cr
   \code{trial_zero_event} \tab The number of trials with at least one arm
   with zero events. When the outcome is continuous, this element is
   omitted. \cr
   \tab \cr
   \code{trial_all_zero_event} \tab The number of trials with zero events in
   all arms. When the outcome is continuous, this element is omitted. \cr
  }

  The four data-frames include \code{interventions_missingness},
  \code{comparisons_missingness}, \code{interventions} and
  \code{comparisons}.
  See 'Value' in \code{\link[rnmamod]{netplot}} that describes these
  data-frames in detail.
}
\description{
This function calculates the necessary elements to describe a
  network, such as the number of interventions, trials, randomised
  participants, and so on. Furthermore, this function provides summary
  statistics of the missing participants and the analysed outcome per
  intervention and observed comparison in a tabulated format. See 'Value' in
  the \code{\link[rnmamod]{netplot}} function for more details.
}
\details{
\code{describe_network} calls the \code{data_preparation} function
  to facilitate the calculations.
}
\seealso{
\code{\link[rnmamod]{netplot}}, \code{\link[rnmamod]{run_model}},
  \code{\link[rnmamod]{data_preparation}}
}
\author{
{Loukia M. Spineli}
}
