% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.metareg_function.R
\name{forestplot_metareg}
\alias{forestplot_metareg}
\title{Comparator-specific forest-plot for network meta-regression}
\usage{
forestplot_metareg(full, reg, compar, cov_value, drug_names)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run_model}}. See 'Value' in
\code{\link{run_model}}.}

\item{reg}{An object of S3 class \code{\link{run_metareg}}. See 'Value' in
\code{\link{run_metareg}}.}

\item{compar}{A character to indicate the comparator intervention. It must
be any name found in \code{drug_names}.}

\item{cov_value}{A vector of two elements in the following order: a number
that corresponds to a value of the covariate considered in
\code{\link{run_metareg}}, and a character object to indicate the name of
the covariate. See also 'Details'.}

\item{drug_names}{A vector of labels with the name of the interventions in
the order they appear in the argument \code{data} of
\code{\link{run_model}}. If the argument \code{drug_names} is not defined,
the order of the interventions as they appear in \code{data} is used,
instead.}
}
\value{
A panel of two forest plots: (1) a forest plot on the estimated
  effect size of comparisons with the selected intervention of the network,
  and (2) a forest plot predicted effect size of comparisons with the
  selected intervention of the network. Both panels illustrate the results
  from network meta-analysis and meta-regression using different colours.
}
\description{
This function illustrates a forest plot of the posterior mean
  and 95\% credible and predictive interval of comparisons with the selected
  intervention of the network.
}
\details{
The y-axis displays all interventions in the network; the selected
  intervention that comprises the \code{compar} is indicated in the plot with
  a homonymous label. The numerical results are displayed above each line.
  Odds ratio and ratio of means are reported in the original scale after
  exponentiation of the logarithmic scale.

  When the covariate is binary, specify in the second element of
  \code{cov_value} the name of the level for which the forest plot will be
  created.

  The interventions are sorted in the descending order of their SUCRA values
  obtain via network meta-analysis.

  \code{forestplot} can be used only for a network of interventions. In the
  case of two interventions, the execution of the function will be stopped
  and an error message will be printed in the R console.
}
\references{
Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries
for presenting results from multiple-treatment meta-analysis: an overview and
tutorial. \emph{J Clin Epidemiol} 2011;\bold{64}(2):163--71.
[\doi{10.1016/j.jclinepi.2010.03.016}]
}
\seealso{
\code{\link{run_model}}, \code{\link{run_metareg}}
}
\author{
{Loukia M. Spineli}
}
