% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld.inconsistency_function.R
\name{kld_inconsistency}
\alias{kld_inconsistency}
\title{Density plots of local inconsistency results and Kullback-Leibler divergence
 when 'rnmamod', 'netmeta' or 'gemtc' R packages are used}
\usage{
kld_inconsistency(node, threshold = 1e-05, drug_names, outcome = NULL)
}
\arguments{
\item{node}{An object of S3 class \code{\link{run_nodesplit}} or class
\code{\link[gemtc:mtc.nodesplit]{mtc.nodesplit}}
(see \href{https://CRAN.R-project.org/package=gemtc}{gemtc}) or
class \code{\link[netmeta:netsplit]{netsplit}} (see
\href{https://CRAN.R-project.org/package=netmeta}{netmeta}).}

\item{threshold}{A positive number indicating the threshold of not concerning
inconsistency, that is, the minimally allowed deviation between the direct
and indirect estimates for a split node that does raise concerns for
material inconsistency. The argument is optional.}

\item{drug_names}{A vector of labels with the name of the interventions in
the order they appear in the argument \code{data}. It is not relevant for
\href{https://CRAN.R-project.org/package=gemtc}{gemtc} and
\href{https://CRAN.R-project.org/package=netmeta}{netmeta}.}

\item{outcome}{Optional argument to describe the effect measure used (the
x-axis of the plots).}
}
\value{
A panel of density plots for each split node sorted in ascending
order of the Kullback-Leibler divergence value. Blue and black lines refer to
the direct and indirect estimates, respectively. The grey segment refers to
the 95\% credible (confidence) interval of the inconsistency parameter, when
\code{\link{run_nodesplit}} (\code{\link[netmeta:netsplit]{netsplit}}) has
been applied, with a darker grey line referring to the point estimate.
When \code{\link[gemtc:mtc.nodesplit]{mtc.nodesplit}} has been employed, the
95\% confidence interval has been approximated using the Bucher's approach
based on the corresponding direct and indirect results.

The Kullback-Leibler divergence value appears at the top left of each plot
in three colours: black, if no threshold has been defined (the default),
green, if the Kullback-Leibler divergence is below the specified
\code{threshold} (not concerning inconsistency) and red, if the
Kullback-Leibler divergence is at least the specified \code{threshold}
(substantial inconsistency).
}
\description{
A panel of density plots on the direct and indirect estimates of the
selected comparisons based on approach for local inconsistency evaluation,
such as back-calculation and node-splitting approaches (Dias et al., 2010;
van Valkenhoef et al., 2016) and loop-specific approach (Bucher et al., 1997)
accompanied by the Kullback-Leibler divergence from the indirect to direct
estimate. The function handles results also from the R-packages
  \href{https://CRAN.R-project.org/package=gemtc}{gemtc} and
  \href{https://CRAN.R-project.org/package=netmeta}{netmeta}.
}
\examples{

data("nma.baker2009")

# Read results from 'run_nodesplit' (using the default arguments)
node <- readRDS(system.file('extdata/node_baker.rds', package = 'rnmamod'))

# The names of the interventions in the order they appear in the dataset
interv_names <- c("placebo", "budesonide", "budesonide plus formoterol",
                  "fluticasone", "fluticasone plus salmeterol",
                  "formoterol", "salmeterol", "tiotropium")

# Apply the function
kld_inconsistency(node = node,
                  threshold = 0.23,
                  drug_names = interv_names,
                  outcome = "Odds ratio (logarithmic scale)")

}
\references{
Bucher HC, Guyatt GH, Griffith LE, Walter SD. The results of direct and
indirect treatment comparisons in meta-analysis of randomized controlled
trials. \emph{J Clin Epidemiol} 1997;\bold{50}(6):683--91.

Dias S, Welton NJ, Caldwell DM, Ades AE. Checking consistency in mixed
treatment comparison meta-analysis.
\emph{Stat Med} 2010;\bold{29}(7-8):932--44.
doi: 10.1002/sim.3767

Kullback S, Leibler RA. On information and sufficiency.
\emph{Ann Math Stat} 1951;\bold{22}(1):79--86. doi: 10.1214/aoms/1177729694

van Valkenhoef G, Dias S, Ades AE, Welton NJ. Automated generation of
node-splitting models for assessment of inconsistency in network
meta-analysis. \emph{Res Synth Methods} 2016;\bold{7}(1):80--93.
doi: 10.1002/jrsm.1167
}
\seealso{
\code{\link[gemtc:mtc.nodesplit]{mtc.nodesplit}},
  \code{\link[netmeta:netsplit]{netsplit}},
  \code{\link{run_nodesplit}}
}
\author{
{Loukia M. Spineli}
}
