% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.diagnostics_function.R
\name{mcmc.diagnostics}
\alias{mcmc.diagnostics}
\title{Markov Chain Monte Carlo Diagnostics}
\usage{
mcmc.diagnostics(net, par)
}
\arguments{
\item{net}{An object of S3 class \code{\link{run.model}} and \code{\link{run.metareg}}. See 'Value' in \code{\link{run.model}}.}

\item{par}{A vector of three character strings that refer to three monitored parameters in \code{jagsfit} which is an object of S3 class \code{\link{run.model}} and \code{\link{run.metareg}}.
These three selected parameters will be considered in the diagnostic plots (see 'Value').}
}
\value{
This function returns a data-frame that contains the Gelman-Rubin convergence diagnostic, R-hat, and convergence status of the following monitored parameters:
\tabular{ll}{
 \code{EM} \tab The estimated summary effect measure.\cr
 \tab \cr
 \code{EM.pred} \tab The predicted summary effect measure.\cr
 \tab \cr
 \code{delta} \tab The estimated trial-specific effect measure.\cr
 \tab \cr
 \code{effectiveneness} \tab The ranking probability of each intervention for every rank.\cr
 \tab \cr
 \code{phi} \tab The informative missingness parameter.\cr
 \tab \cr
 \code{beta} \tab The regression coefficient.\cr
}
\code{mcmc.diagnostics} also uses the \code{\link[mcmcplots]{mcmcplot}} function to create an HTML file with a panel of diagnostic plots (trace, density, and autocorrelation) for each monitored parameter.
}
\description{
This function evaluates whether convergence has been achieved for the monitored parameters of the Bayesian models.
  The Gelman-Rubin convergence diagnostic and relevant diagnostic plots are applied for that purpose.
}
\details{
For each monitored parameter, \code{mcmc.diagnostics} considers the maximum R-hat and compares it with the threshold 1.1: convergence is achieved for the monitored parameter, when the maximum R-hat
  is below that threshold; otherwise, the Markov Chain Monte Carlo algorithm has not converged for that parameter. If the monitored parameter is a vector with the posterior results, there is only one R-hat.
  If the monitored parameter is a matrix of the posterior results, there are as many R-hats as the number of rows for that parameter.
}
\examples{
\dontshow{
load("./data/nma.baker2009.RData")
}

\dontshow{
# Perform a random-effects NMA with consistency equations for the odds ratio (in the logarithmic scale) assuming missing at random for identical, intervention-specific informative missingness odds ratio.
res1 <- run.model(data = nma.baker2009, measure = "OR", model = "RE", assumption = "IDE-ARM", heter.prior = list("halfnormal", 0, 1), mean.misspar = 0, var.misspar = 1, D = 1, n.chains = 3, n.iter = 10000, n.burnin = 1000, n.thin = 1)
}

# Obtain the diagnostic plots and check convergence for all monitored parameters using the R.hat
mcmc.diagnostics(net = res1, par = c("tau", "EM[2,1]", "EM.pred[2,1]"))

}
\references{
Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. \emph{Stat Sci} 1992;\bold{7}:457–-472. [\doi{10.1214/ss/1177011136}]
}
\seealso{
\code{\link[mcmcplots]{mcmcplot}}, \code{\link{run.model}}, \code{\link{run.metareg}}
}
\author{
{Loukia M. Spineli}
}
