% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareg.plot_function.R
\name{metareg.plot}
\alias{metareg.plot}
\title{End-user-ready results: network meta-analysis versus network meta-regression analysis}
\usage{
metareg.plot(full, reg, compar, cov.value, drug.names, save.xls)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{reg}{An object of S3 class \code{\link{run.metareg}}. See 'Value' in \code{\link{run.metareg}}.}

\item{compar}{A character to indicate the comparator intervention. It must be any name found in \code{drug.names}.}

\item{cov.value}{A vector of two elements in the following order: a number that corresponds to a value of the covariate considered in \code{\link{run.metareg}},
and a character object to indicate the name of the covariate. See also 'Details'.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}

\item{save.xls}{Logical to indicate whether to export the tabulated results to an Excel 'xlsx' format (via the \code{\link[writexl]{write_xlsx}} function) to the working directory of the user.
The default is \code{FALSE} (do not export to an Excel format).}
}
\value{
\code{metareg.plot} prints on the R console a message on the most parsimonious model (if any) based on the deviance information criterion (DIC; in red text).
Furthermore, the function returns the following list of elements:
\tabular{ll}{
 \code{Table.estimates} \tab The posterior mean, and 95\% credible interval of the summary effect size for each comparison with the selected intervention under network meta-analysis and meta-regression.\cr
 \tab \cr
 \code{Table.predictions} \tab The posterior mean, and 95\% predictive interval of the summary effect size for each comparison with the selected intervention under network meta-analysis and meta-regression.\cr
 \tab \cr
 \code{Table.model.assessment} \tab The DIC, total residual deviance, number of effective parameters, and the posterior mean and 95% credible interval of between-trial standard deviation (\eqn{\tau}) under each model (Spiegelhalter et al. (2002)).
  When a fixed-effect model has been performed, \code{metareg.plot} does not return results on \eqn{\tau}.\cr
 \tab \cr
 \code{Table.regression.coeffients} \tab The posterior mean and 95\% credible interval of the regression coefficient(s).\cr
 \tab \cr
 \code{Interval.plots} \tab The panel of forest-plots on estimated and predicted effect sizes of comparisons with the selected intervention under network meta-analysis and meta-regression.
  See 'Details' and 'Value' in \code{\link{forestplot.metareg}}.\cr
}
}
\description{
This function illustrates the effect estimates, predictions and regression coefficients of comparisons with a specified comparator and also exports these results in an Excel format.
}
\details{
The DIC of the network meta-analysis model is compared with the DIC of the network meta-regression model. If the difference in DIC exceeds 5, the network meta-regression model is preferred;
  if the difference in DIC is less than -5, the network meta-analysis model is preferred; otherwise, there is little to choose between the compared models.

  When the covariate is binary, specify in the second element of \code{cov.value} the name of the level for which the scatterplot will be created.

  Furthermore, \code{metareg.plot} exports all tabulated results to separate Excel 'xlsx' formats (via the \code{\link[writexl]{write_xlsx}} function) to the working directory of the user.

  \code{metareg.plot} can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed in the R console.
}
\examples{
data("nma.baker2009")

\dontrun{
# Perform a random-effects network meta-analysis
res <- run.model(data = nma.baker2009,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter.prior = list("halfnormal", 0, 1),
                 mean.misspar = c(0, 0),
                 var.misspar = 1,
                 D = 0,
                 n.chains = 3,
                 n.iter = 10000,
                 n.burnin = 1000,
                 n.thin = 1)

# Publicatiom year
pub.year <- c(1996, 1998, 1999, 2000, 2000, 2001, rep(2002, 5), 2003, 2003,
              rep(2005, 4), 2006, 2006, 2007, 2007)

# Perform a random-effects network meta-regression (exchangeable structure)
reg <- run.metareg(full = res,
                   covariate = pub.year,
                   covar.assumption = "exchangeable",
                   n.chains = 3,
                   n.iter = 10000,
                   n.burnin = 1000,
                   n.thin = 1)

# The names of the interventions in the order they appear in the dataset
interv.names <- c("placebo", "budesonide", "budesonide plus formoterol",
                  "fluticasone", "fluticasone plus salmeterol",
                  "formoterol", "salmeterol", "tiotropium")

# Plot the results from the network meta-analysis and meta-regression publication year 2000
# For comparisons with salmeterol
metareg.plot(full = res,
             reg = reg,
             compar = "salmeterol",
             cov.value = c(2000, "publication year"),
             drug.names = interv.names)
}

}
\references{
Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. \emph{J Clin Epidemiol} 2011;\bold{64}(2):163--71. [\doi{10.1016/j.jclinepi.2010.03.016}]

Spiegelhalter DJ, Best NG, Carlin BP, van der Linde A. Bayesian measures of model complexity and fit. \emph{J R Stat Soc B} 2002;\bold{64}:583--616. [\doi{10.1111/1467-9868.00353}]
}
\seealso{
\code{\link{run.model}}, \code{\link{run.metareg}}, \code{\link{forestplot.metareg}}
}
\author{
{Loukia M. Spineli}
}
