% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareg.plot_function.R
\name{metareg.plot}
\alias{metareg.plot}
\title{Plot the results from the meta-regression analysis}
\usage{
metareg.plot(full, metareg, covar.values, drug.names)
}
\description{
Plot the results from the meta-regression analysis
}
