% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.plot_function.R
\name{netplot}
\alias{netplot}
\title{Create the network plot}
\usage{
netplot(data, drug.names, ...)
}
\arguments{
\item{data}{A data-frame of a one-trial-per-row format containing arm-level data of each trial. This format is widely used for BUGS models.
See 'Format' in \code{\link{run.model}} for the specification of the columns.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the \code{data}. If \code{drug.names} is not defined, the order of their interventions
as they appear in the \code{data} is used, instead.}

\item{...}{Additional arguments from the \code{\link[pcnetmeta]{nma.networkplot}} function of the R-package \href{https://CRAN.R-project.org/package=pcnetmeta}{pcnetmeta}.}
}
\value{
A network plot with coloured closed-loops informed by multi-arm trials. Each node refers to the intervention and each link refers to the observed pairwise comparison.
  The edges are proportionally to the number of direct treatment comparisons, unless specified otherwise (see \code{\link[pcnetmeta]{nma.networkplot}} function).
  the node size is weighted by the total number of direct treatment comparisons of the corresponding treatment, unless specified otherwise (see \code{\link[pcnetmeta]{nma.networkplot}} function).
}
\description{
A function to facilitate drawing the network plot via the \code{\link[pcnetmeta]{nma.networkplot}} function
  of the R-package \href{https://CRAN.R-project.org/package=pcnetmeta}{pcnetmeta}. The \code{\link[gemtc]{mtc.data.studyrow}} function
  of the R-package \href{https://CRAN.R-project.org/package=gemtc}{gemtc} is also used to convert \code{data} into a one-arm-per-row format.
}
\examples{
data("nma.baker2009")

head(nma.baker2009)
   t1 t2 t3     y1     y2 y3   sd1   sd2 sd3 m1  m2 m3  n1  n2 n3
#1  1  2 NA -10.70 -13.31 NA  7.64  7.86  NA  1   6 NA 131 253 NA
#2  1  2 NA  -7.19 -12.52 NA 10.57 10.61  NA  3   1 NA 135 137 NA
#3  2  7 NA -15.70 -15.65 NA 10.95 10.95  NA  1  10 NA 175 172 NA
#4  1  2 NA  -3.35  -8.15 NA 14.03 14.08  NA 10   7 NA 132 130 NA
#5  1  2 NA -10.12 -10.80 NA 10.72 10.90  NA  4 138 NA 134 267 NA
#6  1 12 NA  21.60  17.50 NA 11.70 12.15  NA 11  10 NA 100  91 NA

interv.names <- c("placebo", "budesodine", "budesodine plus formoterol", "fluticasone", "fluticasone plus salmeterol",
                  "formoterol", "salmeterol", "tiotropium")
netplot(data = nma.baker2009, drug.names = interv.names)

}
\references{
Lifeng Lin, Jing Zhang, James S. Hodges, Haitao Chu. Performing Arm-Based Network Meta-Analysis in R
with the pcnetmeta Package. \emph{J Stat Softw} 2017;\bold{80}(5): 1--25. <\doi{10.18637/jss.v080.i05}>

Gert van Valkenhoef and Joel Kuiper. gemtc: Network Meta-Analysis Using Bayesian Methods. \emph{R package version 1.0-1}.
2021. \url{https://CRAN.R-project.org/package=gemtc}
}
\seealso{
\code{\link[rnmamod]{run.model}}, \href{https://CRAN.R-project.org/package=pcnetmeta}{pcnetmeta} and \href{https://CRAN.R-project.org/package=gemtc}{gemtc}.
}
\author{
{Loukia M. Spineli}
}
