% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankosucra.plot_function.R
\name{rankosucra.plot}
\alias{rankosucra.plot}
\title{Integrated rankograms and SUCRA curves}
\usage{
rankosucra.plot(full, drug.names)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}
}
\value{
A panel of rankograms with integrated SUCRA curves for each intervention of the network (Salanti et al., 2011). The x-axis of each panel refers to the ranking, and the y-axis refers to the ranking probability in percent.
}
\description{
\code{rankosucra.plot} returns a panel of rankograms with integrated SUCRA curves for each intervention of the network.
}
\details{
Interventions are sorted in the descending order of their SUCRA value.
  The SUCRA value in percent appears on the top left corner of each panel.

  \code{rankosucra.plot} can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed in the R console.
}
\examples{
data("nma.liu2013")

\dontrun{
# Perform a random-effects network meta-analysis
res <- run.model(data = nma.liu2013,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter.prior = list("halfnormal", 0, 1),
                 mean.misspar = 0,
                 var.misspar = 1,
                 D = 1,
                 n.chains = 3,
                 n.iter = 10000,
                 n.burnin = 1000,
                 n.thin = 1)

# The names of the interventions in the order they appear in the dataset
interv.names <- c("placebo", "pramipexole", "serotonin–norepinephrine reuptake inhibitor",
                  "serotonin reuptake inhibitor", "tricyclic antidepressant", "pergolide")

# Create the integrated rankograms and SUCRA curves
rankosucra.plot(full = res, drug.names = interv.names)
}

}
\references{
Salanti G, Ades AE, Ioannidis JP. Graphical methods and numerical summaries for presenting results from multiple-treatment meta-analysis: an overview and tutorial. \emph{J Clin Epidemiol} 2011;\bold{64}(2):163--71. [\doi{10.1016/j.jclinepi.2010.03.016}]
}
\seealso{
\code{\link{run.model}}
}
\author{
{Loukia M. Spineli}, {Chrysostomos Kalyvas}, {Katerina Papadimitropoulou}
}
