% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.metareg_function.R
\name{run.metareg}
\alias{run.metareg}
\title{Pairwise or network meta-regression with missing participant outcome data}
\usage{
run.metareg(
  full,
  covariate,
  covar.assumption,
  n.chains,
  n.iter,
  n.burnin,
  n.thin
)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{covariate}{A numeric vector or a matrix for a trial-specific covariate that is a potential effect modifier. See 'Details'.}

\item{covar.assumption}{Character string indicating the structure of the slope for the intervention by covariate interaction, as described in Cooper et al., (2009).
Set \code{covar.assumption} equal to one of the following: \code{"NO"}, when no meta-regression is performed; otherwise, \code{"exchangeable"} \code{"independent"}, and \code{"common"}.
See the \code{ru.metareg} function.}

\item{n.chains}{Positive integer specifying the number of chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 2.}

\item{n.iter}{Positive integer specifying the number of Markov chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 10000.}

\item{n.burnin}{Positive integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1000.}

\item{n.thin}{Positive integer specifying the thinning rate for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function of the R-package \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
The default argument is 1.}
}
\value{
A list of R2jags outputs on the summaries of the posterior distribution, and the Gelman-Rubin convergence diagnostic (Gelman et al., 1992) of the following monitored parameters for a fixed-effect PMA:
\tabular{ll}{
 \code{EM} \tab The estimated summary effect measure (according to the argument \code{measure}).\cr
 \tab \cr
 \code{dev.o} \tab The deviance contribution of each trial-arm based on the observed outcome.\cr
 \tab \cr
 \code{hat.par} \tab The fitted outcome at each trial-arm.\cr
 \tab \cr
 \code{phi} \tab The informative missingness parameter.\cr
}

For a random-effects PMA, the output additionally includes the following elements:
\tabular{ll}{
 \code{EM.pred} \tab The predicted summary effect measure (according to the argument \code{measure}).\cr
 \tab \cr
 \code{delta} \tab The estimated trial-specific effect measure (according to the argument \code{measure}).
 For a multi-arm trial, we estimate \emph{T-1} effects, where \emph{T} is the number of interventions in the trial.\cr
 \tab \cr
 \code{tau} \tab The between-trial standard deviation.\cr
}

For a random-effects NMA, the output additionally includes:
\tabular{ll}{
 \code{EM.ref} \tab The estimated summary effect measure (according to the argument \code{measure}) of all comparisons with the reference intervention.\cr
 \tab \cr
 \code{pred.ref} \tab The predicted summary effect measure (according to the argument \code{measure}) of all comparisons with the reference intervention.\cr
 \tab \cr
 \code{SUCRA} \tab The surface under the cumulative ranking curve for each intervention.\cr
 \tab \cr
 \code{effectiveneness} \tab The ranking probability of each intervention for every rank.\cr
}
In NMA, \code{EM} and \code{EM.pred} refer to all possible pairwise comparisons of interventions in the network. Furthermore, \code{tau} is typically assumed to be common for all observed comparisons in the network.

Furthermore, the output includes the following elements - the first three resulting from relevant monitored parameters:
\tabular{ll}{
 \code{leverage.o} \tab The leverage for the observed outcome at each trial-arm.\cr
 \tab \cr
 \code{sign.dev.o} \tab The sign of the difference between observed and fitted outcome at each trial-arm.\cr
 \tab \cr
 \code{model.assessment} \tab A data-frame on the measures of model assessment: deviance information criterion, number of effective parameters, and total residual deviance.\cr
 \tab \cr
 \code{jagsfit} \tab An object of S3 class \code{\link[R2jags]{jags}} with the posterior results on all monitored parameters to be used in the \code{mcmc.diagnostics} function.\cr
}
The \code{run.model} function also returns the arguments \code{data}, \code{measure}, \code{model}, \code{assumption}, \code{heter.prior}, \code{mean.misspar}, \code{var.misspar}, and \code{D}
as specified by the user to be considered in other functions of the package.
}
\description{
This function performs a one-stage pairwise or network meta-regression while addressing aggregate binary or continuous missing participant outcome data via the pattern-mixture model.
}
\details{
\code{run.metareg} does not contain the arguments \code{data}, \code{measure}, \code{model}, \code{assumption}, \code{heter.prior}, \code{mean.misspar}, and \code{var.misspar} that are found in \code{run.model}.
  This is to prevent misspecifying the Bayesian model as it would make the comparison of the consistency model with and a covariate meaningless.
  Instead, these arguments are contained in the argument \code{full} of the function. Therefore, the user needs first to apply \code{run.model}, and then use \code{run.UME} (see, 'Examples').

  The model as specified by the arguments of \code{run.model} runs in \code{JAGS} and the progress of the simulation appears in the R console.
  The output of \code{run.model} is used as an S3 object by other functions of the package function to be processed further and provide an end-user-ready output.

  The code of Spineli, (2019) and Spineli et al. (2021) has been extended to incorporate one \emph{study-level covariate} variable following the assumptions of Cooper et al. (2009) for the structure of the intervention by covariate interaction.
  The covariate can be either a numeric vector or a matrix (the number of columns equals the maximum number of arms in the dataset).
}
\examples{
data("nma.baker2009")

\dontrun{
# Perform a random-effects network meta-analysis
res <- run.model(data = nma.baker2009,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter.prior = list("halfnormal", 0, 1),
                 mean.misspar = c(0, 0),
                 var.misspar = 1,
                 D = 1,
                 n.chains = 3,
                 n.iter = 10000,
                 n.burnin = 1000,
                 n.thin = 1)

# Publicatiom year
pub.year <- c(1996, 1998, 1999, 2000, 2000, 2001, rep(2002, 5), 2003, 2003,
              rep(2005, 4), 2006, 2006, 2007, 2007)

# Perform a random-effects network meta-regression (exchangeable structure)
run.metareg(full = res,
            covariate = pub.year,
            covar.assumption = "exchangeable",
            n.chains = 3,
            n.iter = 10000,
            n.burnin = 1000,
            n.thin = 1)
}

}
\references{
Spineli LM, Kalyvas C, Papadimitropoulou K. Continuous(ly) missing outcome data in network meta-analysis: a one-stage pattern-mixture model approach. \emph{Stat Methods Med Res} 2021. [\doi{10.1177/0962280220983544}]

Spineli LM. An empirical comparison of Bayesian modelling strategies for missing binary outcome data in network meta-analysis. \emph{BMC Med Res Methodol} 2019;\bold{19}(1):86. [\doi{10.1186/s12874-019-0731-y}]

Cooper NJ, Sutton AJ, Morris D, Ades AE, Welton NJ. Addressing between-study heterogeneity and inconsistency in mixed treatment comparisons: Application to stroke prevention treatments in individuals with non-rheumatic atrial fibrillation. \emph{Stat Med} 2009;\bold{28}(14):1861--81. [\doi{10.1002/sim.3594}]

Gelman A, Rubin DB. Inference from iterative simulation using multiple sequences. \emph{Stat Sci} 1992;\bold{7}:457--472. [\doi{10.1214/ss/1177011136}]
}
\seealso{
\code{\link{run.model}}, \code{\link[R2jags]{jags}}
}
\author{
{Loukia M. Spineli}
}
