% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.UME_function.R
\name{run.UME}
\alias{run.UME}
\title{A function to perform Bayesian unrelated mean effects model for aggregate binary or continuous outcomes}
\format{
The columns of the data frame \code{data} refer to the following ordered elements for a continuous outcome:
\describe{
 \item{\strong{t}}{An intervention identifier.}
 \item{\strong{y}}{The observed mean value of the outcome.}
 \item{\strong{sd}}{The observed standard deviation of the outcome.}
 \item{\strong{m}}{The number of missing outcome data.}
 \item{\strong{c}}{The number of participants completing the assigned intervention.}
 \item{\strong{na}}{The number of compared interventions.}
}
Apart from \strong{na}, all other elements appear in \code{data} as many times as the maximum number of interventions compared in a trial. See, 'Example'.
}
\usage{
run.UME(
  data,
  measure,
  assumption,
  heter.prior,
  net.ref,
  mean.misspar,
  var.misspar,
  n.chains,
  n.iter,
  n.burnin,
  n.thin
)
}
\arguments{
\item{data}{A data-frame of a one-trial-per-row format containing arm-level data of each trial. This format is widely used for BUGS models. See 'Format' for the specification of the columns.}

\item{measure}{Character string indicating the effect measure with values \code{"OR"}, \code{"MD"}, \code{"SMD"}, or \code{"ROM"}.}

\item{assumption}{Character string indicating the structure of the informative missingness parameter. Set \code{assumption} equal to one of the following: \code{"HIE-COMMON"}, \code{"HIE-TRIAL"}, \code{"HIE-ARM"}, \code{"IDE-COMMON"}, \code{"IDE-TRIAL"}, \code{"IDE-ARM"}, \code{"IND-CORR"}, or \code{"IND-UNCORR"}.}

\item{heter.prior}{A vector of length equal to two with the following values: \code{rep(1, 2)}, \code{rep(2, 2)}, and \code{rep(3, 2)} refers to half-normal distribution with variance 1 or 0.5, and uniform distribution with interval [0, 5], respectively,
for the between-trial standard deviation. To indicate an empirically-based prior distribution for the between-trial variance, the first and second values of the vector should be the mean and precision
of the selected prior distribution. The empirically-based prior distribution for the between-trial variance is applicable only when \code{"OR"} or \code{"SMD"} is considered.}

\item{mean.misspar}{A positive non-zero number for the mean of the normal distribution of the informative missingness parameter.}

\item{var.misspar}{A positive non-zero number for the variance of the normal distribution of the informative missingness parameter.}

\item{n.chains}{Integer specifying the number of chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function.}

\item{n.iter}{Integer specifying the number of Markov chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function.}

\item{n.burnin}{Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function.}

\item{n.thin}{Integer specifying the thinning rate for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function.}
}
\value{
An R2jags output on the summaries of the posterior distribution, and the Gelman–Rubin convergence diagnostic of the following parameters:
\describe{
 \item{\code{EM}}{The effect estimate of all observed comparisons of interventions in the network.}
 \item{\code{dev.o}}{The deviance contribution of each trial-arm based on the observed outcomes.}
 \item{\code{resdev.o}}{The total residual deviance contribution of every trial based on the observed outcomes.}
 \item{\code{totresdev.o}}{The total residual deviance based on the observed outcomes.}
 \item{\code{dev.m}}{The deviance contribution of each trial-arm based on the missing outcomes.}
 \item{\code{resdev.m}}{The total residual deviance contribution of every trial based on the missing outcomes.}
 \item{\code{resdev.m}}{The total residual deviance based on the missing outcomes.}
 \item{\code{tau}}{The between-trial standard deviation assumed to be common for all observed comparisons.}
}
}
\description{
A function to perform Bayesian unrelated mean effects model for aggregate binary or continuous outcomes
}
\examples{
### Show the data (one-trial-per-row format)
(data <- as.data.frame(one.stage.dataset.NMA[[3]]))

### Run a random-effects network meta-analysis with consistency equations for the standardised mean difference
### assuming missing at random for identical, common informative missingness difference of means.
run.UME(data = data, measure = "SMD", assumption = "IDE-COMMON", mean.misspar = 0, var.misspar = 1, n.chains = 3, n.iter = 10000, n.burnin = 1000, n.thin = 1)

}
\references{
Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci. 1992;7:457–472.

\dontshow{load("./data/NMA Dataset Continuous.RData")}
}
\seealso{
\code{\link{R2jags}}
}
