% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.deviance_function.R
\name{scatterplots.dev}
\alias{scatterplots.dev}
\title{Scatterplots on the posterior mean of deviance between consistency and unrelated mean effects (UME) models}
\usage{
scatterplots.dev(full, ume, colour)
}
\description{
Scatterplots on the posterior mean of deviance between consistency and unrelated mean effects (UME) models
}
