% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.meta.plot_function.R
\name{series.meta.plot}
\alias{series.meta.plot}
\title{Forest plot of results: consistency model versus series of pairwise meta-analyses}
\usage{
series.meta.plot(full, meta, drug.names, save.xls)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{meta}{An object of S3 class \code{\link{run.series.meta}}. See 'Value' in \code{\link{run.series.meta}}.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the 
argument \code{drug.names} is not defined, the interventions are ordered as they appear in \code{data}.}

\item{save.xls}{Logical to indicate whether to export the tabulated results in Excel 'xlsx' format at the working directory of the user.
The default is \code{FALSE} (do not export in Excel format).}
}
\value{
\code{series.meta.plot} returns a panel of two forest plots: (1) a forest plot on the posterior mean and 95\% credible interval of the effect of the observed comparisons under the consistency model and the corresponding pairwise meta-analyses,
  and (2) a forest plot on the posterior median and 95\% credible interval of the between-trial standard deviation (\eqn{\tau}) for the observed comparisons. The estimated \eqn{\tau} from the consistency model appears as a rectangle in the forest plot.
  When a fixed-effect model has been fitted, only the forest plot on the estimated effects is shown.

  The R console prints the data-frame with the estimated effects of the observed comparisons under both models and the data-frame on the estimated (\eqn{\tau})s from the pairwise meta-analyses.
  Furthermore, \code{series.meta.plot} exports both data-frames in Excel 'xlsx' format at the working directory of the user.
}
\description{
Facilitates the comparison of the estimated effects and between-trial standard deviation (where applicable) from the network meta-analysis versus the results from the pairwise meta-analyses.
}
\details{
\code{series.meta.plot} can be used only for a network of interventions. 
The user can detect any inconsistencies in the estimated effects from the compared models and explore the gains in their precision by applying the consistency model.
Furthermore, the user can investigate the plausibility of the common between-trial heterogeneity assumption which is typically considered in the consistency model.
}
\examples{
data("nma.dogliotti2014")

\dontrun{
# Perform a random-effects network meta-analysis
res <- run.model(data = nmadogliotti2014,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter.prior = list("halfnormal", 0, 1),
                 mean.misspar = c(0, 0),
                 var.misspar = 1,
                 D = 1,
                 n.chains = 3,
                 n.iter = 10000,
                 n.burnin = 1000,
                 n.thin = 1)

# Run separate random-effects pairwise meta-analyses
meta <- run.series.meta(full = res, n.chains = 3, n.iter = 10000, n.burnin = 1000, n.thin = 1)

# The names of the interventions in the order they appear in the dataset
interv.names <- c("placebo", "aspirin", "aspirin plus clopidogrel", "dabigatran 110 mg",
                  "dabigatran 150 mg", "rivaroxaban", "vitamin K antagonist", "apixaban")

# Plot the results from both models
series.meta.plot(full = res, meta = meta, drug.names = interv.names)
}

}
\seealso{
\code{\link{run.model}}, \code{\link{run.series.meta}}
}
\author{
{Loukia M. Spineli}
}
