% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.meta.plot_function.R
\name{series_meta_plot}
\alias{series_meta_plot}
\title{End-user-ready results for series of pairwise meta-analyses}
\usage{
series_meta_plot(full, meta, drug_names, save_xls)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run_model}}. See 'Value' in
\code{\link{run_model}}.}

\item{meta}{An object of S3 class \code{\link{run_series_meta}}. See 'Value'
in \code{\link{run_series_meta}}.}

\item{drug_names}{A vector of labels with the name of the interventions in
the order they appear in the argument \code{data} of
\code{\link{run_model}}. If \code{drug_names} is not defined,
the order of the interventions as they appear in \code{data} is used,
instead.}

\item{save_xls}{Logical to indicate whether to export the tabulated results
to an 'xlsx' file at the working directory of the user.
The default is \code{FALSE} (do not export).}
}
\value{
The R console prints the data-frame with the estimated summary effect sizes
  and \eqn{\tau}s of comparisons under both models. The comparisons have at
  least two trialsIn the case of a fixed-effect model, the data-frame is
  printed without results on \eqn{\tau}.

  Furthermore, \code{series_meta_plot} exports the data-frame to an 'xlsx'
  file at the working  directory of the user.

  \code{series_meta_plot} returns a panel of two forest plots: (1) a
  forest plot on the posterior mean and 95\% credible interval of the summary
  effect size of the observed comparisons from network meta-analysis and the
  corresponding pairwise meta-analyses, and (2) a forest plot on the
  posterior median and 95\% credible interval of the between-trial standard
  deviation (\eqn{\tau}) for these observed comparisons. The estimated
  \eqn{\tau} from network meta-analysis appears as a rectangle in the forest
  plot. When a fixed-effect model has been fitted, only the forest plot on
  the estimated summary effect sizes is shown.

  The user can detect any inconsistencies in the estimated
  effects from the compared models and explore the gains in precision
  stemming from applying network meta-analysis. Furthermore, the user can
  investigate the plausibility of the common between-trial heterogeneity
  assumption which is typically considered in network meta-analysis.
}
\description{
Facilitates the comparison of the consistency model
  (via \code{\link{run_model}}) with a series of pairwise meta-analyses
  (via \code{\link{run_series_meta}}) regarding the estimated summary effect
  sizes and between-trial standard deviation for comparisons with at least
  two trials.
}
\details{
\code{series_meta_plot} can be used only for a network of
  interventions. Otherwise, the execution of the function will be stopped and
  an error message will be printed on the R console.
}
\examples{
data("nma.dogliotti2014")

\dontrun{
# Perform a random-effects network meta-analysis
res <- run_model(data = nma.dogliotti2014,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter_prior = list("halfnormal", 0, 1),
                 mean_misspar = c(0, 0),
                 var_misspar = 1,
                 D = 1,
                 n_chains = 3,
                 n_iter = 10000,
                 n_burnin = 1000,
                 n_thin = 1)

# Run separate random-effects pairwise meta-analyses
meta <- run_series_meta(full = res,
                        n_chains = 3,
                        n_iter = 10000,
                        n_burnin = 1000,
                        n_thin = 1)

# The names of the interventions in the order they appear in the dataset
interv_names <- c("placebo", "aspirin", "aspirin plus clopidogrel",
                  "dabigatran 110 mg", "dabigatran 150 mg", "rivaroxaban",
                  "vitamin K antagonist", "apixaban")

# Plot the results from both models
series_meta_plot(full = res,
                 meta = meta,
                 drug_names = interv_names)
}
}
\seealso{
\code{\link{run_model}}, \code{\link{run_series_meta}}
}
\author{
{Loukia M. Spineli}
}
