% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UME.plot_function.R
\name{UME.plot}
\alias{UME.plot}
\title{End-user-ready results: consistency model versus unrelated mean effects model}
\usage{
UME.plot(full, ume, drug.names, save.xls)
}
\arguments{
\item{full}{An object of S3 class \code{\link{run.model}}. See 'Value' in \code{\link{run.model}}.}

\item{ume}{An object of S3 class \code{\link{run.UME}}. See 'Value' in \code{\link{run.UME}}.}

\item{drug.names}{A vector of labels with the name of the interventions in the order they appear in the argument \code{data} of \code{\link{run.model}}. If the argument \code{drug.names} is not defined, the order of the interventions
as they appear in \code{data} is used, instead.}

\item{save.xls}{Logical to indicate whether to export the tabulated results to an Excel 'xlsx' format (via the \code{\link[writexl]{write_xlsx}} function) to the working directory of the user.
The default is \code{FALSE} (do not export to an Excel format).}
}
\value{
\code{UME.plot} prints on the R console a message on the most parsimonious model (if any) based on the deviance information criterion (DIC; in red text).
Then, the function returns the following list of elements:
\tabular{ll}{
 \code{Table.effect.size} \tab The posterior mean, posterior standard deviation, and 95\% credible interval of the summary effect size for each pairwise comparison observed in the network under the consistency model and the unrelated mean effects model.\cr
 \tab \cr
 \code{Table.model.assessment} \tab The DIC, number of effective parameters, and total residual deviance under the consistency model and the unrelated mean effects model (Spiegelhalter et al. (2002)).\cr
 \tab \cr
 \code{Table.tau} \tab The posterior median and 95\% credible interval of \eqn{\tau} under the consistency model and the unrelated mean effects model.
  When a fixed-effect model has been performed, \code{UME.plot} does not return this element.\cr
 \tab \cr
 \code{Scatterplots} \tab The scatterplot and the Bland-Altman plot on the posterior mean deviance contribution of the individual data points under the consistency model and the unrelated mean effects model.
  See 'Details' and 'Value' in \code{\link{scatterplots.dev}} and \code{\link{BlandAltman.plot}}.\cr
 \tab \cr
 \code{Levarage.plots} \tab The leverage plot on the posterior mean of deviance of the individual data points under the consistency model and the unrelated mean effects model, separately.
  See 'Details' and 'Value' in \code{\link{leverage.plot}}.\cr
 \tab \cr
 \code{Intervalplots} \tab A panel of interval plots on the summary effect size under the consistency model and the unrelated mean effects model for each pairwise comparison observed in the network.
  See 'Details' and 'Value' in \code{\link{intervalplot.panel.UME}}.\cr
}
}
\description{
This function hosts a toolkit of functions that facilitates the comparison of the consistency model (via \code{run.model}) with the unrelated mean effects model (via \code{run.UME}) regarding the posterior summaries of the summary effect size for
  the pairwise comparisons observed in the network, the between-trial standard deviation (\eqn{\tau}) and model assessment parameters.
}
\details{
The DIC of the consistency model is compared with the DIC of the unrelated mean effects model (Dias et al. (2013)). If the difference in DIC exceeds 5, the unrelated mean effects model is preferred; if the difference in DIC is less than -5,
  the consistency is preferred; otherwise, there is little to choose between the compared models.

  Furthermore, \code{UME.plot} exports \code{EM.both.models} to an Excel 'xlsx' format (via the \code{\link[writexl]{write_xlsx}} function) to the working directory of the user.

  \code{UME.plot} can be used only for a network of interventions. In the case of two interventions, the execution of the function will be stopped and an error message will be printed in the R console.
}
\examples{
data("nma.liu2013")

\dontrun{
# Perform a random-effects network meta-analysis (consistency model)
res <- run.model(data = nma.liu2013,
                 measure = "OR",
                 model = "RE",
                 assumption = "IDE-ARM",
                 heter.prior = list("halfnormal", 0, 1),
                 mean.misspar = c(0, 0),
                 var.misspar = 1,
                 D = 1,
                 n.chains = 3,
                 n.iter = 10000,
                 n.burnin = 1000,
                 n.thin = 1)

# Run random-effects network meta-analysis with node-splitting approach
ume <- run.UME(full = res, n.chains = 3, n.iter = 10000, n.burnin = 1000, n.thin = 1)

# The names of the interventions in the order they appear in the dataset
interv.names <- c("placebo", "pramipexole", "serotonin–norepinephrine reuptake inhibitor",
                  "serotonin reuptake inhibitor", "tricyclic antidepressant", "pergolide")

# Plot the results from the consistency model and the node-splitting approach
UME.plot(full = res, ume = ume, drug.names = interv.names)
}

}
\references{
Spineli LM. A novel framework to evaluate the consistency assumption globally in a network of interventions. \emph{submitted} 2021.

Spineli LM, Kalyvas C, Papadimitropoulou K. Quantifying the robustness of primary analysis results: A case study on missing outcome data in pairwise and network meta-analysis. \emph{Res Synth Methods} 2021;\bold{12}(4):475--490. [\doi{10.1002/jrsm.1478}]

Dias S, Welton NJ, Sutton AJ, Caldwell DM, Lu G, Ades AE. Evidence synthesis for decision making 4: inconsistency in networks of evidence based on randomized controlled trials. \emph{Med Decis Making} 2013a;\bold{33}(5):641--56. [\doi{10.1177/0272989X12455847}]

Spiegelhalter DJ, Best NG, Carlin BP, van der Linde A. Bayesian measures of model complexity and fit. \emph{J R Stat Soc B} 2002;\bold{64}:583--616. [\doi{10.1111/1467-9868.00353}]
}
\seealso{
\code{\link{run.model}}, \code{\link{run.UME}}, \code{\link{BlandAltman.plot}}, \code{link{BlandAltman.plot}}, \code{\link{leverage.plot}}, \code{\link{intervalplot.panel.UME}}
}
\author{
{Loukia M. Spineli}
}
